Compiler.Environment.Importer

Description

This module contains functions for importing module interfaces into environments.

Documentation

importEntries :: [EnvEntry] -> Environment -> Environment

Inserts multiple entries into the given environment and associates them with their original names.

importInterface :: ModuleInterface -> Environment -> Environment

Imports all entries from the given module interface into the given interface.

This function imports entries that are not exported by the given interface as well.