Compiler.Frontend.Haskell.Parser

Description

This module contains functions for parsing Haskell modules and other nodes of the Haskell AST.

We are using the haskell-src-ext package for parsing. This module just provides an interface for the actual parser and configures the parser appropriately.

Documentation

parseHaskell

Arguments

:: (Functor ast, Parseable (ast SrcSpanInfo))

=> String
The name of the Haskell source file.

-> String
The Haskell source code.

-> Reporter (ast SrcSpan)

Parses a node of the Haskell AST.

Modules

parseModule

Arguments

:: String
The name of the Haskell source file.

-> String
The Haskell source code.

-> Reporter (Module SrcSpan)

Parses a Haskell module.

Syntax errors cause a fatal error message to be reported.

parseModuleWithComments

Arguments

:: String
The name of the Haskell source file.

-> String
The Haskell source code.

-> Reporter (Module SrcSpan, [Comment])

Like parseModule but returns the comments in addtion to the AST.

parseModuleFile

Arguments

:: String
The name of the Haskell source file.

-> ReporterIO (Module SrcSpan)

Loads and parses a Haskell module from the file with the given name.

parseModuleFileWithComments

Arguments

:: String
The name of the Haskell source file.

-> ReporterIO (Module SrcSpan, [Comment])

Like parseModuleFile but returns the comments in addtion to the AST.

Declarations

parseDecl

Arguments

:: String
The name of the Haskell source file.

-> String
The Haskell source code.

-> Reporter (Decl SrcSpan)

Parses a Haskell type.

Types

parseType

Arguments

:: String
The name of the Haskell source file.

-> String
The Haskell source code.

-> Reporter (Type SrcSpan)

Parses a Haskell type.

parseTypeSchema

Arguments

:: String
The name of the Haskell source file.

-> String
The Haskell source code.

-> Reporter (Type SrcSpan)

Parses a Haskell type schema.

A type schema is a type with an optional explicit forall quantifier. This requires the ExplicitForAll language extension to be enabled.

Expressions

parseExpr

Arguments

:: String
The name of the Haskell source file.

-> String
The Haskell source code.

-> Reporter (Exp SrcSpan)

Parses a Haskell expression.

Identifiers

parseQName :: String -> Reporter (QName SrcSpan)

Parses an optionally qualified Haskell identifier or symbol of a constructor.

Since there is no Parseable instance for QName, the given string is parsed as a pattern instead. The name of the constructor is extracted from the pattern.