Compiler.Frontend.Haskell.Simplifier

Description

This module contains a converter from the haskell-src-exts AST to the intermediate representation used by the compiler.

The simplifier checks whether a given module complies with our assumptions about the input file format and generates a fatal error message if any Haskell feature is used that is not supported by the compiler.

In some places the simplifier also performs desugaring. For example, infix operators are translated to regular function applications and list literals to the application of the list constructors.

TODO warn user if reserved names like error or undefined are used in declarations or as (type-) variable names.

Documentation

type Simplifier = Converter

The simplifier is a special converter that converts the haskell-src-exts AST to the simplified internal representation for Haskell modules.

During simplification the environment is usually empty (except for fresh variables created by the simplifier).

simplifyModuleWithComments :: Module SrcSpan -> [Comment] -> Simplifier Module

Simplifies a module declaration.

The module must not contain any import declaration or pragmas. The export list may be present but is ignored.

Only regular (non XML) modules are supported.

extractModName :: Module SrcSpan -> Simplifier ModName

Gets the name of the given module.

simplifyDecls :: [Decl SrcSpan] -> Simplifier ([TypeDecl], [TypeSig], [FuncDecl])

Simplifies the given declarations.

simplifyType :: Type SrcSpan -> Simplifier Type

Simplifies the a type expression.

simplifyTypeSchema :: Type SrcSpan -> Simplifier TypeSchema

Simplifies the given type expression and abstracts it to a type schema.

If there is no explicit forall, all type variables that occur in the type expression are abstracted aw. Otherwise, only the specified type arguments are abstracted.

simplifyExpr :: Exp SrcSpan -> Simplifier Expr

Simplifies an expression.