FreeC.Analysis.RecursionAnalysis

Description

This module contains functions for analysising recursive function, e.g. to finding the decreasing argument of a recursive function and to find constant arguments of recursive functions.

Decreasing arguments

type DecArgIndex = Int

Type for the index of a decreasing argument.

identifyDecArgs :: [FuncDecl] -> Converter [DecArgIndex]

Identifies the decreasing arguments of the given mutually recursive function declarations.

Reports a fatal error message, if the decreasing arguments could not be identified.

Constant arguments

data ConstArg

Data type that represents a constant argument shared by multiple mutually recusive functions.

Contains the names and indicies of the corrsponding arguments of the function declarations and a fresh identifier for the Variable sentence that replaces the constant argument.

Constructors

ConstArg

Fields

constArgIdents :: Map QName String

Maps the names of functions that share the constant argument to the names of the corresponding function arguments.

constArgIndicies :: Map QName Int

Maps the names of functions that share the constant argument to the index of the corresponding function argument.

constArgFreshIdent :: String

A fresh identifier for the constant argument.

identifyConstArgs :: [FuncDecl] -> Converter [ConstArg]

Identifies function arguments that can be moved to a Section sentence in Coq.

The call graph of the given function declarations should be strongly connected.