FreeC.Application.Option.Version

Description

This module contains the implementation of the --version command.

The template Haskell language extension is used to include the hash of the Git commit at compile time.

Documentation

putVersionInfo :: IO ()

Prints version information for the compiler, the compiler it has been compiled with and operating system it is running on.

The version string has the following format.

The Free Compiler, version <version> (<commit>, <compiler>, <os>, <arch>)

where

version is the version of the compiler as specified in the Cabal configuration file.

commit is the output of git describe --always --dirty at compile time. The output has the format tag-n-ghash[-dirty] where tag is the name of the latest tag, n is the number of commits since the latest tag, hash is the abbreviated Git commit hash and the suffix -dirty is appended if there are uncommitted changes. If the compiler is not installed from the Git repository, this information is missing.

os is the name of the operating system (e.g., "linux").

arch is the name of the machine architecture (e.g., "x86_64").