FreeC.Application.Options.Descriptors

Description

This module contains descriptors for the command line options that are understood by the command line option parser (see FreeC.Application.Options.Parser).

Documentation

optionDescriptors :: [OptDescr (Options -> Options)]

The command line option descriptors from the GetOpt library for the command line options understood by the command line option parser.