FreeC.Application.Options.Parser

Description

This module contains the command line option parser.

Documentation

parseArgs

Arguments

:: Options
The default options.

-> [String]
The command line arguments.

-> Reporter Options

Parses the command line arguments.

If there are errors when parsing the command line arguments, a fatal error message is reported.

All non-option arguments are considered as input files.

Returns the default options (first argument) if no arguments are specified.

getAndParseArgs :: Options -> ReporterIO Options

Gets the Options for the command line arguments that were passed to the application.

If there are no command line arguments the given default options are returned. Otherwise the given options are modified accordingly.