FreeC.Backend

Description

This module contains the Backend data type and all available backends.

Documentation

data Backend

Data type that represents a backend.

Constructors

Backend

Fields

backendName :: String

The name of the backend.

backendConvertModule :: BackendConverter

The conversion function that converts a module to program text.

backendFileExtension :: String

The file extension associated with the backend.

backendSpecialAction :: Application ()

An action that has to performed by the backend before conversion, e.g. project file creation.

type BackendConverter = Module -> Application String

Type synonym for the conversion function of a Backend.

backends :: Map String Backend

A map of all available backends with the name of those backends as keys.

showBackends :: String

Shows a list of all backends.

defaultBackend :: String

Shows the name of the default backend.