FreeC.Backend.Agda.Base

Description

This module contains the Agda identifiers of types, constructors and functions defined in the Base library that accompanies the compiler.

Library Imports

baseLibName :: Name

The name of the Agda Base library.

generatedLibName :: Name

The name of the Agda library where generated Agda files are placed.

imports :: [Declaration]

Import declaration for the Free and partial modules from the Base Agda base library.

Free Monad

free :: Name

Identifier for the Free monad.

pure :: Name

Identifier for the pure constructor of the Free monad.

shape :: Name

Reserved name for the Shape type variable.

position :: Name

Reserved name for the Pos type variable.

partial :: Name

Reserved name for the Partial type class.

Sized Types

size :: Name

The name of Agda's Size data type.

Size

up :: Name

The Name of the Agda function for larger Size.

↑

Reserved Identifiers

reservedIdents :: [Name]

All Agda identifiers that are reserved for the Base library.

This does only include identifiers without corresponding Haskell name.