FreeC.Backend.Agda.Converter

Description

This module exports functions for generating Agda that uses the Free monad from out intermediate representation.

Documentation

convertModule :: Module -> Converter Declaration

Converts a Haskell module to an Agda declaration.

convertFuncDecl :: FuncDecl -> Converter [Declaration]

Converts the given function declarations. Returns the declarations for the type signature and the definition (TODO).

convertType :: Type -> Converter Expr

Lifts a single type.

convertFunctionType :: [Type] -> Type -> Converter Expr

Functions not defined in terms lambdas (f x = … not f = x -> …) aren't evaluated until fully applied, i.e. cannot be bottom. Therefore interleaving monadic layers isn't needed.

τ₁ -> … -> τₙ -> ρ ↦ τ₁' → … → τₙ' → ρ'

convertConstructorType :: [Type] -> Type -> Converter Expr

Constructors aren't evaluated until fully applied. Furthermore the return type of a constructor for a data type D has to be D and therefore cannot be lifted.

τ₁ -> … -> τₙ -> ρ ↦ τ₁' → … → τₙ' → ρ*

renameAgdaTypeVar :: TypeVarDecl -> Converter Name

Utility function for introducing a new Agda type variable to the current scope.