FreeC.Backend.Agda.Converter.Free

Description

This module contains auxilary functions that help to generate Coq code that uses the Free monad.

Documentation

generatePure :: Expr -> Expr

Applies the pure constructor of the free monad to the given expression.

free :: Expr -> Expr

Lifts a type in the free monad.

applyFreeArgs :: QName -> Expr

Apply the Shape and Pos argument to the given type constructor.

c ↦ c Shape Pos

addFreeArgs :: [Name] -> [Name]

Adds the reserved names for the free args Shape and Pos to a list of names.

freeArgBinder :: [LamBinding]

Binder for the type arguments of the Free monad.

(Shape : Set) (Pos : Shape → Set)