FreeC.Backend.Agda.Converter.Type

Description

Implements the IR to Agda translation, which applies the monadic lifting as described by Abel et al. in "Verifying Haskell Programs Using Constructive Type Theory".

Documentation

convertType :: Type -> Converter Expr

Converts a type from IR to Agda by lifting it into the Free monad.

convertFuncType :: [Type] -> Type -> Converter Expr

Functions not defined in terms lambdas (f x = … not f = x -> …) aren't evaluated until fully applied, i.e. cannot be bottom. Therefore interleaving monadic layers isn't needed.

τ₁ -> … -> τₙ -> ρ ↦ τ₁' → … → τₙ' → ρ'

convertRecFuncType :: Int -> [Type] -> Type -> Converter Expr

convertConType :: [Type] -> Type -> Converter Expr

Constructors aren't evaluated until fully applied. Furthermore the return type of a constructor for a data type D has to be D and therefore cannot be lifted.

τ₁ -> … -> τₙ -> ρ ↦ τ₁' → … → τₙ' → ρ*

convertRecConType :: QName -> [Type] -> Type -> Converter Expr

Converts a constructor type of a recursive data type by lifting it piecewise in the free monad and annotating recursive occurrences of the type with a variable of type Size.