FreeC.Backend.Agda.Keywords

Description

This module exports lists for keywords of the Agda language.

Documentation

agdaKeywords :: [String]

Identifiers that cannot be used in Agda, because they are keywords

This list contains all keywords from the Agda docs https://agda.readthedocs.io/en/v2.6.1/language/lexical-structure.html#keywords-and-special-symbols.