FreeC.Backend.Agda.Syntax

Description

This module exports the original Agda AST to reduce coupling with the Agda libraries.

Identifiers

name :: String -> Name

Creates a (not qualified) Agda variable name from a String.

qname :: [Name] -> Name -> QName

Create a qualified identifier given a local identifier as Name and a list of module Names.

qname' :: Name -> QName

Creates a qualified name using an empty list of module names.

Imports

simpleImport :: QName -> Declaration

Creates an open import declaration for the given qualified name.

open import [modName]

Expressions

intLiteral :: Integer -> Expr

Creates an integer literal.

lambda :: [Name] -> Expr -> Expr

Creates a lambda expression, binding the given names and abstracting the given expression.

λ x y … → [expr]