FreeC.Backend.Coq.Base

Description

This module contains the Coq identifiers of types, constructors and functions defined in the Base library that accompanies the compiler.

Documentation

imports :: Sentence

Import sentence for the Free module from the Base Coq library.

baseLibName :: ModuleIdent

The name of the Base Coq library.

generatedLibName :: ModuleIdent

The name of the Coq library where generated Coq files are placed.

free :: Qualid

The Coq identifier for the Free monad.

freePureCon :: Qualid

The Coq identifier for the pure constructor of the Free monad.

freeImpureCon :: Qualid

The Coq identifier for the impure constructor of the Free monad.

freeBind :: Qualid

The Coq identifier for the >>= operator of the Free monad.

freeArgs :: [(Qualid, Term)]

The names and types of the parameters that must be passed to the Free monad. These parameters are added automatically to every defined type and function.

partialArg :: (Qualid, Term)

The name and type of the Partial instance that must be passed to partial functions.

partialUndefined :: Qualid

The identifier for the error term undefined.

partialError :: Qualid

The identifier for the error term error.

integerScope :: Ident

The scope of integer literals.

stringScope :: Ident

The scope of string literals used in error terms.

reservedIdents :: [Qualid]

All Coq identifiers that are reserved for the Base library.

This does only include identifiers without corresponding Haskell name.