FreeC.Backend.Coq.Converter.Arg

Description

This module contains functions for generating Coq function, constructor and type arguments from our intermediate representation.

Documentation

convertTypeVarDecls

Arguments

:: Explicitness
Whether to generate an explicit or implit binder.

-> [TypeVarDecl]
The type variable declarations.

-> Converter [Binder]

Converts the declarations of type variables in the head of a data type or type synonym declaration to a Coq binder for a set of explicit or implicit type arguments.

E.g. the declaration of the type variable a in data D a = ... is translated to the binder (a : Type). If there are multiple type variable declarations as in data D a b = ... they are grouped into a single binder (a b : Type) because we assume all Haskell type variables to be of kind *.

The first argument controlls whether the generated binders are explicit (e.g. (a : Type)) or implicit (e.g. {a : Type}).

convertArgs

Arguments

:: [VarPat]
The function arguments.

-> Maybe Int
The position of the decreasing argument or Nothing if the function does not decrease on any of its arguments.

-> Converter [Binder]

Converts the given arguments of a function declaration or lambda abstraction to Coq binders.

If the type of an argument is not annotated, it's type will be inferred by Coq.

If the function is recursive (i.e., the second argument is not Nothing), its decreasing argument (given index) is not lifted.

convertArg :: VarPat -> Converter Binder

Converts the argument of a function (a variable pattern) to an explicit Coq binder.

If the variable pattern has a type annotation, the generated binder is annotated with the converted type.

convertPureArg :: VarPat -> Converter Binder

Like convertArg but marks the variable as non-monadic.

If the variable pattern has a type annotation, the generated binder is annotated with the converted type but the type is not lifted to the Maybe monad.

generateArgBinder :: Qualid -> Maybe Term -> Converter Binder

Generates an explicit Coq binder for a function argument with the given name and optional Coq type.

If no type is provided, it will be inferred by Coq.

convertAnonymousArg :: Maybe Type -> Converter (Qualid, Binder)

Converts the argument of an artifically generated function to an explicit Coq binder. A fresh Coq identifier is selected for the argument and returned together with the binder.