FreeC.Backend.Coq.Converter.FuncDecl.Rec.WithHelpers

Description

This module contains functions for converting mutually recursive function declarations by splitting them into one or more recursive helper function, whose decreasing argument is not lifted to the Free monad, and a non-recursive main function.

Documentation

convertRecFuncDeclsWithHelpers :: [FuncDecl] -> Converter [Sentence]

Converts recursive function declarations into recursive helper and non-recursive main functions.

convertRecFuncDeclsWithHelpers' :: [FuncDecl] -> Converter ([Sentence], [Sentence])

Like convertRecFuncDeclsWithHelpers but does return the helper and main functions separately.