FreeC.Backend.Coq.Converter.Module

Description

This module contains functions for converting Haskell modules to Coq.

Documentation

convertModule :: Module -> Converter [Sentence]

Converts a Haskell module to a Gallina sentences.

convertModule' :: Module -> Converter [Sentence]

Like convertModule' but does not apply any compiler passes beforehand.

convertDecls :: [TypeDecl] -> [FuncDecl] -> Converter [Sentence]

Converts the given declarations of a Haskell module.

convertTypeDecls :: [TypeDecl] -> Converter [Sentence]

Converts the given data type or type synonym declarations.

convertImportDecls :: [ImportDecl] -> Converter [Sentence]

Converts the given import declarations to Coq.

convertImportDecl :: ImportDecl -> Converter Sentence

Convert a import declaration.

generateImport :: ModuleIdent -> ModName -> Converter Sentence

Generates a Coq import sentence for the module with the given name from the given library.

Modules from the base library are imported via From Base Require Import sentences and all other modules are also exported.