FreeC.Backend.Coq.Syntax

Description

This module contains smart constructors for nodes of the Coq AST. For convenience the original Coq AST is exported as well.

Comments

comment :: String -> Sentence

Smart constructor for Coq comments.

commentedSentences :: String -> [Sentence] -> [Sentence]

Puts a comment in front of the sentences if there is any sentence.

unpackComment :: Comment -> String

Gets the string from theCoq comment.

Proofs

blankProof :: Proof

An admitted proof that contains only a placeholder text.

Identifiers

ident :: String -> Ident

Smart constructor for unqualified Coq identifiers.

bare :: String -> Qualid

Smart constructor for unqualified Coq identifiers.

qualified :: String -> String -> Qualid

Smart constructor for qualified Coq identifiers.

unpackQualid :: Qualid -> Maybe String

Gets the identifier for the given unqualified Coq identifier. Returns Nothing if the given identifier is qualified.

access :: ModuleIdent -> Ident -> Ident

Smart constructor for combining a module name and an identifier.

Functions

app :: Term -> [Term] -> Term

Smart constructor for the application of a Coq function or (type) constructor.

If the first argument is an application term, the arguments are added to that term. Otherwise a new application term is created.

explicitApp :: Qualid -> [Term] -> Term

Smart constructor for the explicit application of a Coq function or constructor to otherwise inferred type arguments.

If there are no type arguments, no ExplicitApp will be created.

arrows

Arguments

:: [Term]
The types of the function arguments.

-> Term
The return type of the function.

-> Term

Smart constructor for a Coq function type.

fun :: [Qualid] -> [Maybe Term] -> Term -> Term

Smart constructor for the construction of a Coq lambda expression with the given arguments and right-hand side.

The second argument contains the types of the arguments are inferred by Coq.

inferredFun :: [Qualid] -> Term -> Term

Like fun, but all argument types are inferred.

Binders

typedBinder :: Generalizability -> Explicitness -> [Qualid] -> Term -> Binder

Smart constructor for an explicit or implicit typed Coq binder.

typedBinder' :: Generalizability -> Explicitness -> Qualid -> Term -> Binder

Like typedBinder but for a single identifier.

Assumptions

variable :: [Qualid] -> Term -> Sentence

Generates a Variable assumption sentence.

Definition Sentences

definitionSentence

Arguments

:: Qualid
The name of the definition.

-> [Binder]
Binders for the parameters of the definition.

-> Maybe Term
The return type of the definition.

-> Term
The right-hand side of the definition.

-> Sentence

Smart constructor for a Coq definition sentence.

Notation sentences

notationSentence

Arguments

:: NonEmpty NotationToken
The notation to define.

-> Term
The right-hand side of the notation.

-> [SyntaxModifier]
The syntax modifiers of the notation.

-> Sentence

Smart constructor for a Coq notation sentence.

nSymbol :: String -> NotationToken

Smart constructor for a notation token that is a keyword of the notation.

nIdent :: String -> NotationToken

Smart constructor for a notation token that is a variable in the notation.

sModLevel :: Num -> SyntaxModifier

Smart constructor for a parsing level syntax modifier.

sModIdentLevel :: NonEmpty String -> Maybe Num -> SyntaxModifier

Smart constructor for a identifier parsing level syntax modifier.

Types

sortType :: Term

The type of a type variable.

Expressions

string :: String -> Term

Smart constructor for Coq string literals.

equation :: Pattern -> Term -> Equation

Smart constructor for Coq equations for match expressions.

match :: Term -> [Equation] -> Term

Smart constructor for a Coq match expression.

equals :: Term -> Term -> Term

Smart constructor for reflexive equality in Coq.

notEquals :: Term -> Term -> Term

Smart constructor for reflexive inequality in Coq.

conj :: Term -> Term -> Term

Smart constructor for a conjunction in Coq.

disj :: Term -> Term -> Term

Smart constructor for a disjunction in Coq.

Options

setOption :: Maybe Locality -> String -> Maybe (Either Num String) -> Sentence

Smart constructor for a sentence which sets an option or flag.

unsetOption :: Maybe Locality -> String -> Sentence

Smart constructor for a sentence which unsets an option or flag.

Imports

requireImportFrom :: ModuleIdent -> [ModuleIdent] -> Sentence

Creates a From … Require Import … sentence.

requireExportFrom :: ModuleIdent -> [ModuleIdent] -> Sentence

Creates a From … Require Export … sentence.

requireFrom :: ModuleIdent -> [ModuleIdent] -> Sentence

Creates a From … Require … sentence.

moduleImport :: [ModuleIdent] -> Sentence

Creates an Import … sentence.

moduleExport :: [ModuleIdent] -> Sentence

Creates an Export … sentence.