FreeC.Environment

Description

This module contains a data type that encapsulates the state of the compiler. There are also utility functions to modify the state and retrieve information stored in the state.

Environment

data Environment

Data type that encapsulates the state of the converter.

Constructors

Environment

Fields

envAvailableModules :: Map ModName ModuleInterface

Maps names of modules that can be imported to their interface.

envEntries :: Map ScopedName EnvEntry

Maps original names of entries for declarations to the entries.

envDecArgs :: Map QName (Int, String)

Maps Haskell function names to the index and name of their decreasing argument. Contains no entry for non-recursive functions, but there are also entries for functions that are shadowed by local variables.

envFreshIdentCount :: Map String Int

The number of fresh identifiers that were used in the environment with a certain prefix.

emptyEnv :: Environment

An environment that does not even contain any predefined types and functions.

Module information

makeModuleAvailable :: ModuleInterface -> Environment -> Environment

Inserts the interface of a module name into the environment such that it can be imported.

isModuleAvailable :: ModName -> Environment -> Bool

Tests whether the module with the given name can be imported.

lookupAvailableModule :: ModName -> Environment -> Maybe ModuleInterface

Looks up the environment of another module that can be imported.

Inserting entries into the environment

addEntry :: EnvEntry -> Environment -> Environment

Inserts an entry into the given environment and associates it with its original name.

defineDecArg :: QName -> Int -> String -> Environment -> Environment

Stores the index and name of the decreasing argument of a recursive function in the environment.

removeDecArg :: QName -> Environment -> Environment

Removes the index of the decreasing argument of a recursive function from the environment (e.g., because the function has been transformed into a non-recursive function).

Looking up entries from the environment

lookupEntry :: Scope -> QName -> Environment -> Maybe EnvEntry

Looks up the entry with the given original name in the given scope of the given environment.

isFunction :: QName -> Environment -> Bool

Tests whether the given name identifies a function in the given environment.

Returns False if there is no such function.

isVariable :: QName -> Environment -> Bool

Tests whether the given name identifies a local variable in the given environment.

Returns False if there is no such variable.

isPureVar :: QName -> Environment -> Bool

Test whether the variable with the given name is not monadic.

lookupIdent :: Scope -> QName -> Environment -> Maybe Qualid

Looks up the Coq identifier for a Haskell function, (type) constructor or (type) variable with the given name.

Returns Nothing if there is no such function, (type/smart) constructor, constructor or (type) variable with the given name.

lookupSmartIdent :: QName -> Environment -> Maybe Qualid

Looks up the Coq identifier for the smart constructor of the Haskell constructor with the given name.

Returns Nothing if there is no such constructor.

usedIdents :: Environment -> [Qualid]

Gets a list of Coq identifiers for functions, (type/smart) constructors, (type/fresh) variables that were used in the given environment already.

lookupSrcSpan :: Scope -> QName -> Environment -> Maybe SrcSpan

Looks up the location of the declaration with the given name.

lookupTypeArgs :: Scope -> QName -> Environment -> Maybe [TypeVarIdent]

Looks up the type variables used by the type synonym, (smart) constructor or type signature of the function with the given name.

Returns Nothing if there is no such type synonym, function or (smart) constructor with the given name.

lookupTypeArgArity :: Scope -> QName -> Environment -> Maybe Int

Returns the length of the list returned by lookupTypeArgs.

lookupArgTypes :: Scope -> QName -> Environment -> Maybe [Maybe Type]

Looks up the argument and return types of the function or (smart) constructor with the given name.

Returns Nothing if there is no such function or (smart) constructor with the given name.

lookupReturnType :: Scope -> QName -> Environment -> Maybe Type

Looks up the return type of the function or (smart) constructor with the given name.

Returns Nothing if there is no such function or (smart) constructor with the given name or the return type is not known.

lookupTypeSchema :: Scope -> QName -> Environment -> Maybe TypeSchema

Gets the type schema of the variable, function or constructor with the given name.

lookupArity :: Scope -> QName -> Environment -> Maybe Int

Looks up the number of arguments expected by the Haskell function or smart constructor with the given name.

Returns Nothing if there is no such function or (smart) constructor with the given name.

lookupTypeSynonym :: QName -> Environment -> Maybe ([TypeVarIdent], Type)

Looks up the type the type synonym with the given name is associated with.

Returns Nothing if there is no such type synonym.

needsFreeArgs :: QName -> Environment -> Bool

Tests whether the function with the given name needs the arguments of the Free monad.

Returns False if there is no such function.

isPartial :: QName -> Environment -> Bool

Tests whether the function with the given name is partial.

Returns False if there is no such function.

lookupDecArg :: QName -> Environment -> Maybe (Int, String)

Looks up the index and name of the decreasing argument of the recursive function with the given name.

Returns Nothing if there is no such recursive function.

lookupDecArgIndex :: QName -> Environment -> Maybe Int

Like lookupDecArg but returns the decreasing argument's index only.

lookupDecArgIdent :: QName -> Environment -> Maybe String

Like lookupDecArg but returns the decreasing argument's name only.