FreeC.Environment.Fresh

Description

This module contains functions for generating fresh identifiers.

Fresh identifiers are identifiers that have been introduced artificially into the Haskell or Coq AST. They are guaranteed not to conflict with any other valid identifier.

Documentation

freshArgPrefix :: String

The prefix to use for artificially introduced variables of type a.

freshFuncPrefix :: String

The prefix to use for artificially introduced variables of type a -> b.

freshBoolPrefix :: String

The prefix to use for artificially introduced variables of type Bool.

freshTypeVarPrefix :: String

The prefix to use for aritifcially introduced type variables of kind *.

freshTypeArgPrefix :: String

The prefix to use for artificially introduced type variables of kind * that are passed as a type argument to a function.

freshHaskellIdent :: String -> Converter String

Gets the next fresh Haskell identifier from the current environment.

All fresh identifiers contain an at-sign (See internalIdentChar). This ensures that they cannot be confused with actual Haskell identifiers. The corresponding Coq identifier contains an underscore instead (see FreeC.Environment.Renamer). The at-sign (or underscore) is preceded by the given prefix and followed by an incrementing number (staring at 0).

If the given prefix is a fresh identifier itself (i.e. contains an at-sign), the prefix of that identifier is used instead.

The freshly generated identifier is not inserted into the environment, i.e. it can still be used to create a declaration.

freshHaskellName :: Name -> Converter Name

Applies freshHaskellIdent to the given name.

Fails if the given name is a symbol.

freshHaskellQName :: QName -> Converter QName

Like freshHaskellName but preserves the module name of qualified names.

freshTypeVar :: Converter Type

Generates a fresh Haskell type variable.

freshCoqIdent :: String -> Converter String

Gets the next fresh Haskell identifier from the current environment and renames it such that it can be used in Coq.

The freshly generated identifier is not inserted into the environment, i.e. it can still be used to create a declaration.

freshCoqQualid :: String -> Converter Qualid

Like freshCoqIdent but the resulting Coq identifier is wrapped in a Term.