FreeC.Environment.LookupOrFail

Description

This module contains functions to lookup entries of the Environment that (in contrast to the functions defined in FreeC.Environment) report a fatal error message when there is no such entry.

Documentation

lookupEntryOrFail :: SrcSpan -> Scope -> QName -> Converter EnvEntry

Looks up an entry of the environment with the given name or reports a fatal error message if the identifier has not been defined or the name is ambigious.

If an error is reported, it points to the given source span.

lookupIdentOrFail

Arguments

:: SrcSpan
The source location where the identifier is requested.

-> Scope
The scope to look the identifier up in.

-> QName
The Haskell identifier to look up.

-> Converter Qualid

Looks up the Coq identifier for a Haskell function, (type) constructor or (type) variable with the given name or reports a fatal error message if the identifier has not been defined.

If an error is reported, it points to the given source span.

lookupSmartIdentOrFail

Arguments

:: SrcSpan
The source location where the identifier is requested.

-> QName
The Haskell identifier to look up.

-> Converter Qualid

Looks up the Coq identifier of a smart constructor of the Haskell data constructr with the given name or reports a fatal error message if there is no such constructor.

If an error is reported, it points to the given source span.