FreeC.Environment.ModuleInterface.Encoder

Description

This module contains functions for encoding ModuleInterfaces in JSON and writing them to .json files.

Encoding module interfaces as TOML files is not supported, since TOML is intended for human maintained configuration files (e.g., the module interface of the Prelude) only.

See FreeC.Environment.ModuleInterface.Decoder for more information on the interface file format.

Documentation

writeModuleInterface :: (MonadIO r, MonadReporter r) => FilePath -> ModuleInterface -> r ()

Serializes a module interface and writes it to a .json file.

Orphan instances

ToJSON Qualid

Instance details

Methods

toJSON :: Qualid -> Value

toEncoding :: Qualid -> Encoding

toJSONList :: [Qualid] -> Value

toEncodingList :: [Qualid] -> Encoding

ToJSON QName

Instance details

Methods

toJSON :: QName -> Value

toEncoding :: QName -> Encoding

toJSONList :: [QName] -> Value

toEncodingList :: [QName] -> Encoding

ToJSON Type

Instance details

Methods

toJSON :: Type -> Value

toEncoding :: Type -> Encoding

toJSONList :: [Type] -> Value

toEncodingList :: [Type] -> Encoding

ToJSON ModuleInterface
Serializes a ModuleInterface.

Instance details

Methods

toJSON :: ModuleInterface -> Value

toEncoding :: ModuleInterface -> Encoding

toJSONList :: [ModuleInterface] -> Value

toEncodingList :: [ModuleInterface] -> Encoding