FreeC.Frontend.Haskell.PatternMatching

Description

This module provides an interface to the pattern matching compiler and case complition library by Malte Clement https://git.informatik.uni-kiel.de/stu204333/placc-thesis. We are using a slightly adapted version of the library located at https://github.com/FreeProving/haskell-src-transformations.

Documentation

transformPatternMatching :: Module SrcSpan -> Converter (Module SrcSpan)

Applies the pattern matching transformation, guard elimination and case completion.

Since the pattern matching compiler library does not support source spans, location information is removed during the transformation.