FreeC.Frontend.Haskell.SrcSpanConverter

Description

This module contains type class instances for converting source spans from the haskell-src-exts package to the source spans of the intermediate representation of the compiler.

Documentation

class ConvertibleSrcSpan ss where

Type class for source spans from other packages that can be converted to SrcSpans for pretty printing of messages.

Methods

convertSrcSpan :: SrcFileMap -> ss -> SrcSpan

Converts the given third party source span to a SrcSpan by attaching the corresponding line of source code.

Orphan instances

ConvertibleSrcSpan SrcLoc
Converts a SrcLoc by creating a zero width source span and applying the conversion for SrcSpans.

Instance details

Methods

convertSrcSpan :: SrcFileMap -> SrcLoc -> SrcSpan

ConvertibleSrcSpan SrcSpan
Directly converts a SrcSpan to a SrcSpan by looking up the corresponding line of code in the provided map.

Instance details

Methods

convertSrcSpan :: SrcFileMap -> SrcSpan -> SrcSpan0

ConvertibleSrcSpan SrcSpanInfo
Converts a SrcSpanInfo by removing additional information and applying the conversion for SrcSpans.

Instance details

Methods

convertSrcSpan :: SrcFileMap -> SrcSpanInfo -> SrcSpan