FreeC.Frontend.IR.ParserTests

Description

This module contains tests for FreeC.Frontend.IR.Parser.

Documentation

shouldParse :: (Eq a, Parseable a, Show a) => String -> a -> Expectation

Sets the expectation that the Parseable instance accepts the given input and produces the given output.

shouldParseModule :: [String] -> Module -> Expectation

Like shouldParse for modules.

Automatically concatenates the lines of the input module.

shouldBeParseError :: (Parseable a, Show a) => Reporter a -> Expectation

Sets the expectation that the given parser reports a fatal message.

testIRParser :: Spec

Test group for FreeC.Frontend.IR.Parser tests.

testNameParser :: Spec

Test group for Parseable instance of Name.

testQNameParser :: Spec

Test group for Parseable instance of QName.

testTypeParser :: Spec

Test group for Parseable instance of Type.

testTypeSchemeParser :: Spec

Test group for Parseable instance of TypeScheme.

testSynTypeDeclParser :: Spec

Test group for Parseable instance of type synonym declarations.

testDataDeclParser :: Spec

Test group for Parseable instance of data type declarations.

testTypeSigParser :: Spec

Test group for Parseable instance of TypeSig.

testExprParser :: Spec

Test group for Parseable instance of Expr.

testExprTypeParser :: Spec

Test group for Parseable instance of expressions with type annotations.

testConExprParser :: Spec

Test group for Parseable instance of constructor expressions.

testVarExprParser :: Spec

Test group for Parseable instance of variable expressions.

testAppExprParser :: Spec

Test group for Parseable instance of function application expressions.

testLambdaExprParser :: Spec

Test group for Parseable instance of lambda abstractions.

testIfExprParser :: Spec

Test group for Parseable instance of if expressions.

testCaseExprParser :: Spec

Test group for Parseable instance of case expressions.

testEffectParser :: Spec

Test group for Parseable instance of effects.

testTypeAppExprParser :: Spec

Test group for Parseable instance of visible type applications.

testIntLiteralParser :: Spec

Test group for Parseable instance of integer literal expressions.

testFuncDeclParser :: Spec

Test group for Parseable instance of function declarations.

testImportDeclParser :: Spec

Test group for Parseable instance of ImportDecl.

testModuleParser :: Spec

Test group for Parseable instance of Module.