FreeC.Frontend.IR.Scanner

Description

This module contains a scanner for the intermediate language that takes the source code and converts it into a token stream.

We convert the source code to a token stream such that FreeC.Frontend.IR.Parser does not have to handle white space explicitly.

Documentation

data TokenWithPos

A Token and its position in the source code.

Constructors

TokenWithPos

Fields

getTokenPos :: SourcePos

getToken :: Token

scan :: MonadReporter r => SrcFile -> r [TokenWithPos]

Converts the given IR source code to a stream of IR tokens.

Reports a fatal error if there are unknown tokens.