FreeC.IR.DependencyGraph

Description

This module contains functions to construct dependency graphs of declarations and modules of the intermediate language. A dependency graph is a directed graph whose nodes are labeled with declarations. There is an edge from node A to B if the declaration of A depends on the declaration of B (i.e. in Coq B has to be defined before A or both have to be declared in the same sentence). The module dependency graph contains an edge from node A to B if the module of A contains an import declaration for the module of B.

The dependency graph does only contain defined top-level declarations (i.e. there are no nodes for imported or built-in data types or operations and there are no nodes for local variables such as function parameters or variable patterns). However, the entries of the dependency graph contain keys for predefined functions (but not local variables) as well as the special functions error and undefined that are used in error terms.

We distinguish between the type and value dependency graph. This is because function declarations and type declarations live in separate scopes but we want to avoid name conflicts. Since we assume all type declarations to precede function declarations in the generated Coq code, this separation of the dependency graphs should not be a problem. For the same reason, the value dependency graph does not include nodes for constructors (as always, the keys of used constructors are still present).

The construction of a dependency graph does not fail even if there are are undefined identifiers. The dependency graph does not know about the environment, it only knows which (type) constructors and (type) variables occur freely in the declarations. Most importantly, that means that it does not know about the module system. Thus, dependency analysis should not be performed before all references have been resolved to their original (qualified) identifiers.

For debugging purposes, dependency graphs can be converted to the DOT format such that they can be visualized using Graphviz (See https://www.graphviz.org/).

Dependency Graph

type DGKey = QName

Every node of the dependency graph is uniquely identified by a key. We use their qualified original name to identify declarations.

type DGEntry node = (node, DGKey, [DGKey])

Every node (i.e., declaration) in a dependency graph is associated with a unique key (its qualified original name) and a list of keys that identify the nodes this node depends on (adjacency list).

data DependencyGraph node

A dependency graph is a directed graph whose nodes are declarations (Usually TypeDecl or FuncDecl). There is an edge from node A to node B if the declaration of A depends on B.

Nodes are identified by their qualified original name (See DGKey). Internally, nodes are identified by a number (See Vertex).

In addition to the actual Graph that stores the adjacency matrix of the internal identifiers, this data type contains functions to convert between the internal and high level representation.

Constructors

DependencyGraph

Fields

dgGraph :: Graph

The actual graph.

dgGetEntry :: Vertex -> DGEntry node

Gets an entry for a vertex of the graph.

dgGetVertex :: DGKey -> Maybe Vertex

Gets the vertex of a node with the given key.

Getters

dgEntries :: DependencyGraph node -> [DGEntry node]

Gets the entries of the given dependency graph.

Predicates

dependsDirectlyOn

Arguments

:: DependencyGraph node
The dependency graph.

-> DGKey
The key of the first node.

-> DGKey
The key of the second node.

-> Bool

Tests whether two nodes (identified by the given keys) of the dependency graph depend on each other directly (i.e., there is an edge from the first node to the second node).

Returns False if one of the nodes does not exist.

Constructors

typeDependencyGraph :: (HasRefs node, HasDeclIdent node) => [node] -> DependencyGraph node

Creates the type dependency graph for a list of nodes.

valueDependencyGraph :: (HasRefs node, HasDeclIdent node) => [node] -> DependencyGraph node

Creates the dependency graph for a list of function declarations.

moduleDependencyGraph :: [ModuleOf decls] -> DependencyGraph (ModuleOf decls)

Creates the dependency graph for the given modules.

Strongly Connected Components

data DependencyComponent decl

A strongly connected component of the dependency graph.

All declarations that mutually depend on each other are in the same strongly connected component.

The only difference to SCC decl is that the constructors have been renamed to be more explanatory in the context of dependency analysis.

Constructors

NonRecursive decl
A single non-recursive declaration.

Recursive [decl]
A list of mutually recursive declarations.

unwrapComponent :: DependencyComponent decl -> [decl]

Gets the declarations wrapped by the given strongly connected component.

Constructors

dependencyComponents :: DependencyGraph a -> [DependencyComponent a]

Computes the strongly connected components of the given dependency graph.

Each strongly connected component corresponds either to a set of mutually recursive declarations or a single non-recursive declaration.

The returned list of strongly connected components is in reverse topological order, i.e. if any declaration in a strongly connected component B depends on a declaration in a strongly connected component A, then A precedes B in the returned list.

If two components do not depend on each other, they are in reverse alphabetical order.

typeDependencyComponents :: (HasRefs node, HasDeclIdent node) => [node] -> [DependencyComponent node]

Combines the construction of the type dependency graph with the computation of strongly connected components.

valueDependencyComponents :: (HasRefs node, HasDeclIdent node) => [node] -> [DependencyComponent node]

Combines the construction of the value dependency graph with the computation of strongly connected components.

groupModules :: [ModuleOf decls] -> [DependencyComponent (ModuleOf decls)]

Combines the construction of the dependency graph for the given modules (See moduleDependencyGraph) with the computation of strongly connected components.

Since cyclic module dependencies are not allowed, all DependencyComponents in the returned list should be NonRecursive. Otherwise, there is a module dependency error.

Manipulating

mapComponent :: ([decl] -> [decl']) -> DependencyComponent decl -> DependencyComponent decl'

Applies the given function to the declarations in the given strongly connected component of the dependency graph.

In case of a NonRecursive component, the function is given a singleton list. The given function must not change the number of declarations in the component.

mapComponentM :: MonadFail m => ([decl] -> m [decl']) -> DependencyComponent decl -> m (DependencyComponent decl')

Monadic version of mapComponent.

There must be a MonadFail instance since this function fails if the given function changes the number of declarations in a non-recursive strongly connected component.

mapComponentM_ :: MonadFail m => ([decl] -> m a) -> DependencyComponent decl -> m ()

Like mapComponentM but discards the result.