FreeC.IR.Reference

Description

This module contains functions to extract the names of (type) constructors and variables that are referenced by AST nodes such as expressions and type expressions.

These functions are used to construct the dependency graph and to find free (type) variables in (type) expressions.

References

data Ref

Wrapper that is used to remember whether a name refers to a variable or constructor.

The wrapped names are ScopedNames such that we can use the same function to collect the type- and value-level references.

Constructors

VarRef

Fields

unwrapRef :: ScopedName

ConRef

Fields

unwrapRef :: ScopedName

refScope :: Ref -> Scope

Unwraps the given reference and discards the name.

refName :: Ref -> QName

Unwraps the given reference and discards the scope information.

isVarRef :: Ref -> Bool

Tests whether the given reference refers to a variable or type variable.

isConRef :: Ref -> Bool

Tests whether the given reference refers to a constructor or type constructor.

isTypeRef :: Ref -> Bool

Tests whether the given reference refers to a type-level entry.

isValueRef :: Ref -> Bool

Tests whether the given reference refers to a value-level entry.

Finding references

class HasRefs a

Type class for AST nodes that contain references to (type) variables and constructors.

Minimal complete definition

refSet

refs :: HasRefs a => a -> [Ref]

Gets all references to variables, constructors, type variables and type constructors in the given node as they occur from left to right.

typeRefs :: HasRefs a => a -> [QName]

Gets the names of all type variables and type constructors the given node refers to.

valueRefs :: HasRefs a => a -> [QName]

gets the names of all variable and constructors the given node refers to.

Free type variables

freeTypeVars :: HasRefs a => a -> [TypeVarIdent]

The type variables that occur freely in the given node from left to right.

freeTypeVarSet :: HasRefs a => a -> Set TypeVarIdent

The type variables that occur freely in the given node.

Free variables

freeVars :: HasRefs a => a -> [QName]

The variables that occur freely in the given node from left to right.

freeVarSet :: HasRefs a => a -> Set QName

The variables that occur freely in the given node.