FreeC.IR.Similar

Description

This module contains a definition for a custom equality of IR AST nodes. Two AST nodes are called similar if they are equal modulo renaming of bound variables. The source location information is also ignored.

Definition Renaming

A Renaming is a mapping { x₁ ↦ y₁, …, xₙ ↦ yₙ } from variables x₁, …, xₙ in one AST to variables y₁, …, yₙ in another AST. Note that variables and type variables are considered distict even if they have the same name.

Definition: Free Variables in Renaming

A variable x is leftFree in a Renaming Γ if there is no variable y such that x ↦ y ∈ Γ. Analogously a variable y is rightFree in a Renaming Γ if there is no variable x such that x ↦ y ∈ Γ. We write z ∈ free(Γ) if z is both leftFree and rightFree.

Notation

If two nodes n and m are similar under a Renaming Γ we write Γ ⊢ n ≈ m. Whether two nodes are equal or not can be determined from the inference system described in the comments for the Similar instances below.

Documentation

class Similar node

A type class for AST nodes that can be tested for similarity.

Minimal complete definition

similar'

similar :: Similar node => node -> node -> Bool

Tests whether two AST nodes are similar, i.e., if they are equal except for renaming of bound variables.

notSimilar :: Similar node => node -> node -> Bool

The negation of similar.