FreeC.IR.SrcSpan

Description

This module contains the definition of a source span data type that is used by the AST of the intermediate language to store for every node the corresponding portion of the source code.

Source spans can be pretty printed and are used by the error reporting engine of the compiler to point the user to the relevant piece of code when an error occurs.

Source Files

data SrcFile

Data type that contains the name and contents of source files.

The contents of the source file are stored as a string for parsing and as a list of lines for error messages (i.e., to construct SrcSpans).

Constructors

SrcFile

Fields

srcFileName :: FilePath

The name of the file.

srcFileContents :: String

The contents of the file.

srcFileLines :: [String]

The lines of srcFileContents.

type SrcFileMap = [(FilePath, SrcFile)]

Type for a map that associates source files with their filename.

mkSrcFile :: FilePath -> String -> SrcFile

Smart constructor for SrcFile that automatically splits the file contents into lines.

mkSrcFileMap :: [SrcFile] -> SrcFileMap

Smart constructor for SrcFileMap for the given SrcFiles.

Source Spans

data SrcSpan

Describes the portion of the source code that caused a message to be reported.

In contrast to the source spans provided by the haskell-src-exts package this source span provides access to the lines of code that contain the source span.

Constructors

SrcSpan

Fields

srcSpanFilename :: String

The name of the file.

srcSpanStartLine :: Int

The number of the first spanned line.

srcSpanStartColumn :: Int

The offset within the first line.

srcSpanEndLine :: Int

The number of the last spanned line.

srcSpanEndColumn :: Int

The offset within the last line.

srcSpanCodeLines :: [String]

The source code of the spanned lines.

NoSrcSpan
Indicates that no location information is available.

FileSpan
Points to an unknown location in the given file.

Fields

srcSpanFilename :: String

The name of the file.

Predicates

hasSrcSpanFilename :: SrcSpan -> Bool

Tests whether the given SrcSpan contains filename information (i.e., there is a field srcSpanFilename).

hasSourceCode :: SrcSpan -> Bool

Tests whether the given source span has attached source code.

spansMultipleLines :: SrcSpan -> Bool

Tests whether the given source span spans multiple lines.

Conversion

class ConvertableSrcSpan ss where

Type class for source spans from other packages that can be converted to SrcSpans for pretty printing of messages.

Methods

convertSrcSpan :: ss -> SrcSpan

Converts the given third party source span to a SrcSpan by attaching the corresponding line of source code.

convertSrcSpanWithCode :: ConvertableSrcSpan ss => SrcFileMap -> ss -> SrcSpan

Like convertSrcSpan but also adds source code using addSourceCode.

addSourceCode :: SrcFileMap -> SrcSpan -> SrcSpan

Adds source code to the given source span if it does not have source code already.