FreeC.IR.Strip

Description

This module contains functions for removing information from the AST of the intermediate language.

These functions are useful to simplify tests or make the output of pretty instances actually pretty by removing irrelevant information.

Removing Expression Type Annotations

class StripExprType node where

Type class for AST nodes from which expression type annotations can be removed.

Methods

stripExprType :: node -> node