FreeC.IR.Subterm

Description

This module contains data types and functions for working with subterms of expressions and type expressions.

There are also functions for finding the names and types of variables that are bound at a given position in an expression.

Direct Children

childTerms :: Subterm a => a -> [a]

Gets the child nodes of the given AST node.

replaceChildTerms :: Subterm a => a -> [a] -> Maybe a

Replaces the child nodes of the given AST node.

Positions

newtype Pos

Describes a position of a subterm within a Haskell expression.

Constructors

Pos [Int]

rootPos :: Pos

Position of the root expression.

consPos :: Int -> Pos -> Pos

Extends a position inside of a child node to a position inside of a parent node.

If pos is the position of a subterm s of an expression or type expression tᵢ, then consPos i pos is the position of the subterm s of a term C t₁ ... tᵢ ... tₙ.

unConsPos :: Pos -> Maybe (Int, Pos)

Inverse function of consPos.

Returns Nothing if the given position is the rootPos.

parentPos :: Pos -> Maybe Pos

Gets the parent position or Nothing if the given position is the root position.

ancestorPos :: Pos -> [Pos]

Gets the positions of all ancestors of the the given position including the position itself.

allPos :: Subterm a => a -> [Pos]

Gets all valid positions of subterms within the given Haskell expression.

above :: Pos -> Pos -> Bool

Tests whether a position is above another one.

below :: Pos -> Pos -> Bool

Tests whether a position is below another one.

leftOf :: Pos -> Pos -> Bool

Tests whether a position is left of another one.

rightOf :: Pos -> Pos -> Bool

Tests whether a position is right of another one.

Subterms

selectSubterm :: Subterm a => a -> Pos -> Maybe a

Selects a subterm of the given expression at the specified position.

Returns Nothing if there is no such subterm.

selectSubterm' :: Subterm a => a -> Pos -> a

Like selectSubterm but throws an error if there is no such subterm.

replaceSubterm

Arguments

:: Subterm a

=> a
The (type) expression whose subterm to replace.

-> Pos
The position of the subterm.

-> a
The (type) expression to replace the subterm with.

-> Maybe a

Replaces a subterm of the given expression or type expression at the specified position with another expression.

Returns Nothing if there is no such subterm.

replaceSubterm' :: Subterm a => a -> Pos -> a -> a

Like replaceSubterm but throws an error if there is no such subterm.

replaceSubterms :: Subterm a => a -> [(Pos, a)] -> Maybe a

Replaces all subterms at the given positions with other (type) expressions.

Returns Nothing if any of the subterms could not be replaced

replaceSubterms' :: Subterm a => a -> [(Pos, a)] -> a

Like replaceSubterms but throws an error if any of the subterms could not be replaced.

Searching for Subterms

findSubtermPos :: Subterm a => (a -> Bool) -> a -> [Pos]

Gets a list of positions for subterms of the given expression that satisfy the provided predicate.

findSubtermWithPos :: Subterm a => (a -> Pos -> Bool) -> a -> [(a, Pos)]

Like findSubtermPos but the predicate gets the position as an additional argument and also returns the subterm.

findSubterms :: Subterm a => (a -> Bool) -> a -> [a]

Gets a list of subterms of the given expression that satisfy the provided predicate.

findFirstSubterm :: Subterm a => (a -> Bool) -> a -> Maybe a

Gets the first subterm of the given expression that satisfies the provided predicate.

Return Nothing if there is no such subterm.

Replacing Subterms

mapSubterms :: Subterm a => (a -> a) -> a -> a

Applies the given function to all subterms of the given node.

The subterms of the returned expression are replaced recursively.

mapSubtermsM :: (Subterm a, Monad m) => (a -> m a) -> a -> m a

Bound Variables

boundVarsOf :: Expr -> Int -> Set QName

Gets the names of variables that are bound by the given expression in its subterm at the given index.

For example, a case-expression

case e { … ; Cᵢ x₁ … xₙ -> eᵢ ; … }

with subterms [e, e₁, …, eₙ] binds the variables x₁ … xₙ in it's i+1th subterm eᵢ but no variables are bound in e (i.e., if i = 0).

Returns an empty map if the expression does not have such a subterm.

boundVarsWithTypeOf :: Expr -> Int -> Map QName (Maybe Type)

Like boundVarsOf but also returns the annotated type of the variable pattern.

Returns an empty map if the expression does not have such a subterm.

boundVarsAt :: Expr -> Pos -> Set QName

Gets the names of variables that are bound by lambda abstractions or variable patterns in case-expressions at the given position of an expression.

Returns the empty set if the position is invalid.

boundVarsWithTypeAt :: Expr -> Pos -> Map QName (Maybe Type)

Like boundVarsAt but also returns the annotated type of the variable pattern.

Returns an empty map if the position is invalid.