FreeC.IR.SubtermTests

Description

This module contains tests for FreeC.IR.Subterm.

Documentation

validTestPos :: Expr -> Gen Pos

Creates a generator for valid test positions for the given expression.

invalidTestPos :: Expr -> Gen Pos

Creates a generator for invalid test positions for the given expression (i.e. positions that do not identify a subterm of the given expression).

testPos :: Expr -> Gen (Pos, Bool)

Creates a generator for test positions for the given expression.

The Bool indicates whether the position is valid or not.

testSubterm :: Spec

Test group for FreeC.IR.Subterm tests.