FreeC.IR.Syntax.Expr

Description

This module contains the definition of expressions for our intermediate language.

Documentation

data Expr

An expression.

Constructors

Con
A constructor.

Fields

exprSrcSpan :: SrcSpan

exprConName :: ConName

exprTypeSchema :: Maybe TypeSchema

Var
A function or local variable.

Fields

exprSrcSpan :: SrcSpan

exprVarName :: VarName

exprTypeSchema :: Maybe TypeSchema

App
Function or constructor application.

Fields

exprSrcSpan :: SrcSpan

exprAppLhr :: Expr

exprAppRhs :: Expr

exprTypeSchema :: Maybe TypeSchema

TypeAppExpr
Visible type application.

Fields

exprSrcSpan :: SrcSpan

exprTypeAppLhs :: Expr

exprTypeAppRhs :: Type

exprTypeSchema :: Maybe TypeSchema

If
if expression.

Fields

exprSrcSpan :: SrcSpan

ifExprCond :: Expr

ifExprThen :: Expr

ifExprElse :: Expr

exprTypeSchema :: Maybe TypeSchema

Case
case expression.

Fields

exprSrcSpan :: SrcSpan

caseExprScrutinee :: Expr

caseExprAlts :: [Alt]

exprTypeSchema :: Maybe TypeSchema

Undefined
Error term undefined.

Fields

exprSrcSpan :: SrcSpan

exprTypeSchema :: Maybe TypeSchema

ErrorExpr
Error term error "message".

Fields

exprSrcSpan :: SrcSpan

errorExprMsg :: String

exprTypeSchema :: Maybe TypeSchema

IntLiteral
An integer literal.

Fields

exprSrcSpan :: SrcSpan

intLiteralValue :: Integer

exprTypeSchema :: Maybe TypeSchema

Lambda
A lambda abstraction.

Fields

exprSrcSpan :: SrcSpan

lambdaExprArgs :: [VarPat]

lambdaEprRhs :: Expr

exprTypeSchema :: Maybe TypeSchema

exprType :: Expr -> Maybe Type

Gets the type annotation of the given expression but discards the forall.

Type annotations quantify their type variables usually only if they are as expression type signatures. The type annotations generated during type inference never quantify their type arguments.

untypedCon :: SrcSpan -> ConName -> Expr

Smart constructor for Con without the last argument.

untypedVar :: SrcSpan -> ConName -> Expr

Smart constructor for Var without the last argument.

untypedApp :: SrcSpan -> Expr -> Expr -> Expr

Smart constructor for app without the last argument.

The type annotation is inferred from the callee's type annotation. If it is annotated with a function type, the created expression is annotated with the function type's result type.

app :: SrcSpan -> Expr -> [Expr] -> Expr

Creates an expression for applying the given expression to the provided arguments.

The given source span is inserted into the generated function application expressions.

If the given expression's type is annotated with a function type, all generated application nodes are annotated with the corresponding result types. If no more argument types can be split off, the types of the remaining arguments are not annotated.

varApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> VarName
The name of the function to apply.

-> [Expr]
The arguments to pass to the function.

-> Expr

Creates an expression for applying the function with the given name.

The given source span is inserted into the generated function reference and every generated function application.

Since the type of the variable with the given name is not known, no type annotations will be generated.

conApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> ConName
The name of the constructor to apply.

-> [Expr]
The arguments to pass to the constructor.

-> Expr

Creates a data constructor application expression.

The given source span is inserted into the generated constructor reference and every generated constructor application.

Since the type of the constructor with the given name is not known, no type annotations will be generated.

visibleTypeApp :: SrcSpan -> Expr -> [Type] -> Expr

Creates an expression for passing the type arguments of a function or constructor explicitly.

The given source span is inserted into every generated visible type application node.

If the given expression's type is annotated, all generated visible type application nodes are annotated with the same type.

prettyExprPred :: Int -> Expr -> Doc

Pretty prints an expression and adds parenthesis if necessary.

The first argument indicates the precedence of the surrounding context. * 0 - Top level. No parenthesis are necessary. * 1 - Parenthesis are needed around if, case and lambda expressions. * 2 - Parenthesis are also needed around function applications.

data Alt

One alternative of a case expression.

Constructors

Alt

Fields

altSrcSpan :: SrcSpan

altConPat :: ConPat

altVarPats :: [VarPat]

altRhs :: Expr

data ConPat

A constructor pattern used in an alternative of a case expression.

The main purpose of this data type is to add location information to a ConName.

Constructors

ConPat

Fields

conPatSrcSpan :: SrcSpan

conPatName :: ConName

conPatToExpr :: ConPat -> Expr

Converts a constructor pattern to a constructor expression.

data VarPat

A variable pattern used as an argument to a function, lambda abstraction or constructor pattern.

The variable pattern can optionally have a type signature.

Constructors

VarPat

Fields

varPatSrcSpan :: SrcSpan

varPatIdent :: String

varPatType :: Maybe Type

varPatName :: VarPat -> Name

Gets the name of the given variable pattern.

varPatQName :: VarPat -> QName

Gets the qualified name of the given variable pattern.

varPatToExpr :: VarPat -> Expr

Converts a variable pattern to a variable expression.

toVarPat :: String -> VarPat

Converts the given identifier to a variable pattern without type annotation.