FreeC.IR.Syntax.FuncDecl

Description

This module contains the definition of expressions for function declarations and type signatures for our intermediate language.

Documentation

data TypeSig

A type signature of one or more function declarations.

Constructors

TypeSig

Fields

typeSigSrcSpan :: SrcSpan

typeSigDeclIdents :: [DeclIdent]

typeSigTypeScheme :: TypeScheme

data FuncDecl

A function declaration.

Constructors

FuncDecl

Fields

funcDeclSrcSpan :: SrcSpan

A source span that spans the entire function declaration.

funcDeclIdent :: DeclIdent

The name of the function.

funcDeclTypeArgs :: [TypeVarDecl]

The type arguments of the function.

funcDeclArgs :: [VarPat]

The function's argument patterns.

funcDeclReturnType :: Maybe Type

The return type of the function.

funcDeclRhs :: Expr

The right-hand side of the function declaration.

funcDeclQName :: FuncDecl -> QName

Gets the qualified name of the given function declaration.

funcDeclName :: FuncDecl -> Name

Gets the unqualified name of the given function declaration.

funcDeclType :: FuncDecl -> Maybe Type

Gets the type of the given function declaration or Nothing if at least one of the argument or return type is not annotated.

In contrast to funcDeclTypeScheme the function's type arguments are not abstracted away.

funcDeclTypeScheme :: FuncDecl -> Maybe TypeScheme

Gets the type scheme of the given function declaration or Nothing if at least one of the argument or the return type is not annotated.