FreeC.IR.Syntax.Module

Description

This module contains the definition of modules of our intermediate language.

Documentation

data ModuleOf contents

A module declaration that contains declarations of type contents.

This type is used for example to represent partially transformed modules where only the name, imports and pragmas of the module are transformed into the intermediate representation.

Constructors

ModuleOf

Fields

modSrcSpan :: SrcSpan

modName :: ModName

modImports :: [ImportDecl]

modPragmas :: [Pragma]

modContents :: contents

modWithContents :: (contents -> contents') -> ModuleOf contents -> ModuleOf contents'

Applies the given function to the contents of the given module.

type Module = ModuleOf [TopLevelDecl]

A module declaration.

data ImportDecl

An import declaration.

Constructors

ImportDecl

Fields

importSrcSpan :: SrcSpan

importName :: ModName

data TopLevelDecl

Declarations that can occur on top-level of a Module declaration.

Constructors

TopLevelTypeDecl

Fields

topLevelTypeDecl :: TypeDecl

TopLevelTypeSig

Fields

topLevelTypeSig :: TypeSig

TopLevelFuncDecl

Fields

topLevelFuncDecl :: FuncDecl

isTopLevelTypeDecl :: TopLevelDecl -> Bool

Tests whether the given top-level declaration is a TypeDecl.

isTopLevelTypeSig :: TopLevelDecl -> Bool

Tests whether the given top-level declaration is a TypeSig.

isTopLevelFuncDecl :: TopLevelDecl -> Bool

Tests whether the given top-level declaration is a FuncDecl.

modTypeDecls :: Module -> [TypeDecl]

Gets the type declarations of the given module.

modWithTypeDecls :: [TypeDecl] -> Module -> Module

Sets the type declarations of the given module.

All other type declarations are discarded.

modTypeSigs :: Module -> [TypeSig]

Gets the type signatures of the given module.

modWithTypeSigs :: [TypeSig] -> Module -> Module

Sets the type signatures of the given module.

All other type signatures are discarded.

modFuncDecls :: Module -> [FuncDecl]

Gets the function declarations of the given module.

modWithFuncDecls :: [FuncDecl] -> Module -> Module

Sets the function declarations of the given module.

All other function declarations are discarded.