FreeC.IR.Syntax.Name

Description

This module contains the definition of names of our intermediate language.

Documentation

data Name

An identifier or a symbolic name.

The constructors of this type do not contain source spans because Names are intended to be comparable. They are used as keys to identify nodes of the dependency graph for example.

Constructors

Ident String
An identifier, e.g. Ident "f" for a function f.

Symbol String
A symbolic name, e.g. Symbol "+" for (+).

identFromName :: Name -> Maybe String

Extracts an identifier from a name. Returns Nothing if the given name is a symbol and not an identifier.

data QName

A qualifiable Name.

Constructors

Qual ModName Name
A qualified Name.

UnQual Name
An unqualified Name.

nameFromQName :: QName -> Name

Extracts the name of a qualifiable name.

identFromQName :: QName -> Maybe String

Extracts an identifier from a qualifiable name.

toQual :: ModName -> QName -> QName

Converts a qualifiable name to a name that is qualified with the given module name.

toUnQual :: QName -> QName

Converts a qualifiable name to an unqualified name.

type TypeVarIdent = String

The name of a type variable.

type ModName = String

The name of a module.

type VarName = QName

The name of a function or build-in operator used in prefix notation, e.g. f x y or (+) n m

type ConName = QName

The name of an constructor used in prefix notation, e.g. (:) x xs.

type TypeConName = QName

The name of a type or type constructor, e.g. Integer or [] a

data DeclIdent

The name of a top-level declaration including location information.

Constructors

DeclIdent

Fields

declIdentSrcSpan :: SrcSpan

declIdentName :: QName

internalIdentChar :: Char

The character that is used to mark internal identifiers.

This is used to generate fresh identifiers that don't conflict with user defined identifiers.

isInternalIdent :: String -> Bool

Tests whether the given identifier was generated for internal use only (i.e., contains internalIdentChar).

isInternalName :: Name -> Bool

Tests whether the given name was generated for internal use only (i.e., it is an identifier that matches isInternalIdent).

isInternalQName :: QName -> Bool

Tests whether the given qualifiable name was generated for internal use only (i.e., the qualified name is internal according to isInternalName).