FreeC.IR.Syntax.Pragma

Description

This module contains the definition of comments and pragmas for our intermediate language.

Documentation

data Comment

A comment.

Comments are collected during parsing. However, the final AST contains no comments. Pragmas (see DecArgPragma) are extracted from comments by the front end.

Constructors

BlockComment
A multi-line comment (i.e., {- ... -}).

Fields

commentSrcSpan :: SrcSpan

commentText :: String

LineComment
A single-line comment (i.e., -- ...).

Fields

commentSrcSpan :: SrcSpan

commentText :: String

customPragmaPrefix :: String

All custom pragmas of the compiler start with FreeC.

data Pragma

Data type for custom {-# FreeC ... #-} pragmas.

Constructors

DecArgPragma
A {-# FreeC function DECREASES ON argument #-} or {-# FreeC function DECREASES ON ARGUMENT index #-} pragma.

Fields

pragmaSrcSpan :: SrcSpan

decArgPragmaFuncName :: QName

decArgPragmaArg :: Either String Int

prettyPragma :: Doc -> Doc

Pretty prints a custom {-# FreeC ... #-} pragma with the given contents.