FreeC.IR.Syntax.Type

Description

This module contains the definition of type expressions of our intermediate language.

Documentation

data Type

A type expression.

Built-in types are represented by applications of their type constructors. E.g. the type [a] is represented as TypeApp (TypeCon "[]") (TypeVar "a"). The only exception to this rule is the function type a -> b. It is represented directly as FuncType (TypeVar "a") (TypeVar "b"). The syntax (->) a b is not supported at the moment. This is due to the special role of functions during the lifting.

Constructors

TypeVar
A type variable.

Fields

typeSrcSpan :: SrcSpan

typeVarIdent :: TypeVarIdent

TypeCon
A type constructor.

Fields

typeSrcSpan :: SrcSpan

typeConName :: TypeConName

TypeApp
A type constructor application.

Fields

typeSrcSpan :: SrcSpan

typeAppLhs :: Type

typeAppRhs :: Type

FuncType
A function type.

Fields

typeSrcSpan :: SrcSpan

funcTypeArg :: Type

funcTypeRes :: Type

typeApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> Type
The partially applied type constructor.

-> [Type]
The type arguments to pass to the type constructor.

-> Type

Creates a type constructor application type.

The given source span is inserted into the generated type constructor and every generated type constructor application.

typeConApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> TypeConName
The name of the type constructor to apply.

-> [Type]
The type arguments to pass to the type constructor.

-> Type

Creates a type constructor application type for the constructor with the given name.

The given source span is inserted into the generated type constructor and every generated type constructor application.

funcType :: SrcSpan -> [Type] -> Type -> Type

Creates a function type with the given argument and return types.

splitFuncType :: Type -> Int -> ([Type], Type)

Splits the type of a function or constructor with the given arity into the argument and return types.

This is basically the inverse of funcType.

prettyTypePred :: Int -> Type -> Doc

Pretty prints a type and adds parentheses if necessary.

The first argument indicates the precedence of the surrounding context. * 0 - Top level. No parentheses are necessary. * 1 - Parentheses are needed around function types. * 2 - Parentheses are also needed around type constructor applications.