FreeC.IR.Syntax.TypeDecl

Description

This module contains the definition of data type and type synonym declarations of our intermediate language.

Documentation

data TypeDecl

A data type or type synonym declaration.

Constructors

DataDecl

Fields

typeDeclSrcSpan :: SrcSpan

typeDeclIdent :: DeclIdent

typeDeclArgs :: [TypeVarDecl]

dataDeclCons :: [ConDecl]

TypeSynDecl

Fields

typeDeclSrcSpan :: SrcSpan

typeDeclIdent :: DeclIdent

typeDeclArgs :: [TypeVarDecl]

typeSynDeclRhs :: Type

typeDeclQName :: TypeDecl -> QName

Gets the qualified name of the given type declaration.

typeDeclName :: TypeDecl -> Name

Gets the unqualified name of the given type declaration.

data ConDecl

A constructor declaration.

Constructors

ConDecl

Fields

conDeclSrcSpan :: SrcSpan

conDeclIdent :: DeclIdent

conDeclFields :: [Type]

conDeclQName :: ConDecl -> QName

Gets the qualified name of the given constructor declaration.

conDeclName :: ConDecl -> Name

Gets the unqualified name of the given constructor declaration.