FreeC.IR.Syntax.TypeVarDecl

Description

This module contains the definition of type variable declarations of our intermediate language.

Documentation

data TypeVarDecl

The name of a type variable declaration in the head of a data type or type synonym declaration including location information.

Constructors

TypeVarDecl

Fields

typeVarDeclSrcSpan :: SrcSpan

typeVarDeclIdent :: String

typeVarDeclToType :: TypeVarDecl -> Type

Converts the declaration of a type variable to a type.

typeVarDeclName :: TypeVarDecl -> Name

Gets the name of the type variable declared by the given type variable declaration.

typeVarDeclQName :: TypeVarDecl -> QName

Gets the unqualified name of the type variable declared by the given type variable declaration.