FreeC.LiftedIR.Converter.Type

Description

Implements the IR to lifted IR translation, which applies the monadic lifting as described by Abel et al. in "Verifying Haskell Programs Using Constructive Type Theory".

Documentation

liftFuncArgTypes

Arguments

:: Maybe Int
Index of the decreasing argument for recursive functions.

-> [VarPat]
The argument types.

-> [Type]

Converts the argument types of a function.

liftConArgType :: QName -> Type -> Type

Converts a constructor argument using convertType.

All occurrences of the constructed type are marked as structurally smaller.

liftVarPatType :: VarPat -> Maybe Type

Lifts the type of an VarPat. If the argument is strict the type itself isn't lifted into the Free monad.

Translations

liftType :: Type -> Type

Converts a type from IR to lifted IR by lifting it into the Free monad.

This corresponds to the dagger translation for monotypes as described by Abel et al.

τ' = Free τ*

liftType' :: Type -> Type

Lifts a type from IR to lifted IR by adding the free arguments to constructors and lifting function types in the Free monad.

This corresponds to the star translation for monotypes as described by Abel et al.