FreeC.LiftedIR.Syntax.Expr

Description

This module contains the definition of expressions for our lifted intermediate language.

Documentation

data Expr

An expression.

Constructors

Con
A constructor.

Fields

exprSrcSpan :: SrcSpan

exprConName :: ConName

exprType :: Type

Var
A function or local variable.

Fields

exprSrcSpan :: SrcSpan

exprVarName :: VarName

exprType :: Type

App
Function or constructor application.

Fields

exprSrcSpan :: SrcSpan

exprAppFunc :: Expr

exprAppTypeArgs :: [Type]

Visible type applications.

exprEffects :: [Effect]

Effect set.

exprAppArgs :: [Expr]

Applied arguments.

exprType :: Type

If
if expression.

Fields

exprSrcSpan :: SrcSpan

ifExprCond :: Expr

ifExprThen :: Expr

ifExprElse :: Expr

exprType :: Type

Case
case expression.

Fields

exprSrcSpan :: SrcSpan

caseExprScrutinee :: Expr

caseExprAlts :: [Alt]

exprType :: Type

Undefined
Error term undefined.

Fields

exprSrcSpan :: SrcSpan

exprType :: Type

ErrorExpr
Error term error "message".

Fields

exprSrcSpan :: SrcSpan

errorExprMsg :: String

exprType :: Type

IntLiteral
An integer literal.

Fields

exprSrcSpan :: SrcSpan

intLiteralValue :: Integer

exprType :: Type

Lambda
A lambda abstraction.

Fields

exprSrcSpan :: SrcSpan

lambdaExprArgs :: [VarPat]

lambdaEprRhs :: Expr

exprType :: Type

Pure
The pure constructor of the Free monad.

Fields

exprSrcSpan :: SrcSpan

exprPureArg :: Expr

The value that is lifted into the Free monad.

exprType :: Type

Bind
The bind operator for the free monad.

Fields

exprSrcSpan :: SrcSpan

exprBindArg :: Expr

The left-hand side argument of >>=.

exprBindCont :: Expr

The right-hand side argument of >>=.

exprType :: Type

data Alt

One alternative of a case expression.

Constructors

Alt

Fields

altSrcSpan :: SrcSpan

altConPat :: ConPat

altVarPats :: [VarPat]

altRhs :: Expr

data ConPat

A constructor pattern used in an alternative of a case expression.

The main purpose of this data type is to add location information to a ConName.

Constructors

ConPat

Fields

conPatSrcSpan :: SrcSpan

conPatName :: ConName

data VarPat

A variable pattern used as an argument to a function, lambda abstraction or constructor pattern.

The variable pattern can optionally have a type signature.

Constructors

VarPat

Fields

varPatSrcSpan :: SrcSpan

varPatIdent :: String

varPatType :: Maybe Type