FreeC.LiftedIR.Syntax.Name

Description

This module reexports the data types for names from FreeC.IR.Syntax.Name.

Documentation

data Name

An identifier or a symbolic name.

The constructors of this type do not contain source spans because Names are intended to be comparable. They are used as keys to identify nodes of the dependency graph for example.

Constructors

Ident String
An identifier, e.g. Ident "f" for a function f.

Symbol String
A symbolic name, e.g. Symbol "+" for (+).

data QName

A qualifiable Name.

Constructors

Qual ModName Name
A qualified Name.

UnQual Name
An unqualified Name.

Value Level Names

type ConName = QName

The name of a constructor used in prefix notation, e.g. (:) x xs.

type VarName = QName

The name of a function or built-in operator used in prefix notation, e.g. f x y or (+) n m

Type Level Names

type TypeVarIdent = String

The name of a type variable.

type TypeConName = QName

The name of a type or type constructor, e.g. Integer or [] a