FreeC.LiftedIR.Syntax.Type

Description

This module contains the definition of type expressions of our lifted intermediate language.

Documentation

data Type

A type expression.

Types are represented as fully applied type constructors or type variables of kind *. In contrast to the Type from the intermediate representation, Free is modelled explicitly.

Constructors

TypeVar
A type variable.

Fields

typeSrcSpan :: SrcSpan

typeVarIdent :: TypeVarIdent

TypeCon
A fully applied n-ary type constructor application.

Fields

typeSrcSpan :: SrcSpan

typeConName :: TypeConName

typeConArgs :: [Type]

typeIsDec :: Bool

Marks this type as a decreasing element of a type signature.

E.g., recursive constructor arguments or decreasing function argument. This field is used by the Agda back end to place sized type annotations.

FuncType
A function type.

Fields

typeSrcSpan :: SrcSpan

funcTypeArg :: Type

funcTypeRes :: Type

FreeTypeCon
A type constructor for the free monad.

Fields

typeSrcSpan :: SrcSpan

typeFreeArg :: Type