FreeC.Monad.Application

Description

This module contains the state monad used by the compiler's command line interface to pass command line options around implicitly.

State Monad

data Application a

A state monad used by the compiler application to pass the command line options implicitly.

The entire application is lifted to the ConverterIO monad.

runApp :: Application a -> IO a

Runs the compiler application.

reportApp :: Application a -> Application a

Runs the given application and prints the reported messages.

Accessing and Modifying State

getOpts :: Application Options

Gets the options of the application.

inOpts :: (Options -> a) -> Application a

Gets a specific component of the the application's options using the given function to extract the value from the Options.

putOpts :: Options -> Application ()

Sets the options of the application.

modifyOpts :: (Options -> Options) -> Application ()

Modifies the options of the application.

modifyOpts' :: (Options -> (a, Options)) -> Application a

Gets a specific component and modifies the options of the application.

Lifting Other Monads

liftReporter :: MonadReporter r => Reporter a -> r a

Promotes a reporter to r.

liftReporterIO :: ReporterIO a -> Application a

Promotes a ReporterIO to an application that produces the same result and ignores the application's options.

liftConverter :: MonadConverter m => Converter a -> m a

liftConverterIO :: ConverterIO a -> Application a

Promotes a ConverterIO to an application that produces the same result and ignores the application's options.