FreeC.Monad.Class.Testable

Description

This module contains a type class for monads that can be used in tests.

Documentation

class Monad m => MonadTestable m err | m -> err

Type class for monads which can be used in tests.

Minimal complete definition

shouldReturnWith', shouldFailWith'

Expecting values

shouldReturn :: (Eq a, Show a, MonadTestable m err) => m a -> a -> Expectation

Sets the expectation that the given computation successfully returns the given value.

shouldReturnWith :: (Show a, MonadTestable m err) => m a -> (a -> Expectation) -> Expectation

Sets the expectation returned by the given function for the value returned by the given computation.

Expecting success

shouldSucceed :: (Show a, MonadTestable m err) => m a -> Expectation

Sets the expectation that the given computation does not fail.

shouldSucceedWith :: MonadTestable m err => m Expectation -> Expectation

Sets the expectation that the given computation successfully produces an expectation and that expectation holds.

Expecting failures

shouldFail :: (Show a, MonadTestable m err) => m a -> Expectation

Sets the expectation that the given computation fails without returning a value.

shouldFailWith :: (Show a, MonadTestable m err) => m a -> (err -> Expectation) -> Expectation

Sets the expectation returned by the given function for the error that was produced by the given computation instead of a value.

QuickCheck

shouldReturnProperty :: (MonadTestable m err, Testable prop) => m prop -> Property

Sets the expectation that the given computation returns a testable QuickCheck property and returns a property that is satisfied if and only if.