FreeC.Monad.Reporter

Description

This module contains the definition of a monad that is used by the compiler to report error messages, warnings and hints to the user without throwing an exception or performing IO actions.

During execution the Reporter monad collects all reported messages internally. Additionally the monad holds the result of the computation. The computation can be interrupted without returning a result by reporting a fatal error message.

The ReporterT monad transformer is used to implement ReporterIO which simplifies combining IO actions with error reporting.

This module also provides functions for pretty printing the collected error messages in a similar way to how the GHC prints error messages.

Messages

data Message

A message reported by the compiler.

Constructors

Message SrcSpan Severity String

data Severity

The severity of a message reported by the compiler.

Constructors

Internal

Error

Warning

Info

Debug

Reporter Monad

type Reporter = ReporterT Identity

A monad that collects the messages reported by the compiler and contains an optional value that is present only if the compiler did not encounter a fatal error.

This type behaves like (Maybe a, [Message]).

runReporter :: Reporter a -> (Maybe a, [Message])

Runs the given reporter and returns the produced value as well as all reported messages. If a fatal message has been reported the produced value is Nothing.

evalReporter :: Reporter a -> Maybe a

Like runReporter but discards the reported messages.

Reporter Monad Transformer

data ReporterT m a

A reporter monad parameterized by the inner monad m.

runReporterT :: Monad m => ReporterT m a -> m (Maybe a, [Message])

Runs the given reporter and returns the produced value as well as all reported messages. If a fatal message has been reported the produced value is Nothing. The result is wrapped in the inner monad.

lift :: (MonadTrans t, Monad m) => m a -> t m a

Lift a computation from the argument monad to the constructed monad.

hoist :: (Hoistable t, Monad m) => t Identity a -> t m a

Lifts the transformed identity monad to any transformed monad.

unhoist :: (UnHoistable t, Monad m) => t m a -> m (t Identity a)

Undoes hoist.

Reporting Messages

class Monad r => MonadReporter r where

Type class for all monads within which Messages can be reported.

Methods

liftReporter :: Reporter a -> r a

Promotes a reporter to r.

report :: MonadReporter r => Message -> r ()

Creates a successful reporter that reports the given message.

reportFatal :: MonadReporter r => Message -> r a

Creates a reporter that fails with the given message.

Reporting IO Errors

type ReporterIO = ReporterT IO

A reporter with an IO action as its inner monad.

liftIO :: MonadIO m => IO a -> m a

Lift a computation from the IO monad.

reportIOError :: MonadReporter r => IOError -> r a

Reports the given IO error as a fatal error with no location information.

Handling Messages and Reporter Results

isFatal :: Reporter a -> Bool

Tests whether a fatal error was reported to the given reporter.

messages :: Reporter a -> [Message]

Gets the messages reported to the given reporter.

Handling Reported Messages

reportTo :: MonadIO m => Handle -> ReporterT m a -> m (Maybe a)

Runs the given reporter and prints all reported messages to the provided file handle.

If the inner monad of the reporter is an IO action, the IO action will be executed before the messages are printed to the file handle. To run an IO action after the messages have been reported, the reporter needs to return the IO action (e.g. Reporter (IO ()) instead of ReporterIO ()). It is possible to combine both approaches (i.e. run an IO action before the messages are printed and another action afterwards) by using ReporterIO (IO ()). In the latter case this function returns a value of type IO (IO ()). Thus an additional join is needed: join (reportTo h reporter).

reportToOrExit :: MonadIO m => Handle -> ReporterT m a -> m a

Runs the given reporter, prints all reported messages to stderr and exits the application if a fatal message has been reported.

See reportTo for usage information.