FreeC.Pass.EtaConversionPassTests

Description

This module contains tests for FreeC.Pass.EtaConversionPass.

Documentation

shouldEtaConvertTopLevel :: String -> String -> Converter Expectation

Parses the given expressions, applies the eta conversion pass and sets the expectation that the resulting expression is similar to the expected output.

shouldEtaConvert :: String -> String -> Converter Expectation

testEtaConversionPass :: Spec

Test group for etaConversionPass tests.