FreeC.Pass.PartialityAnalysisPass

Description

This module contains a compiler pass that updates the partiality information of function declaration environment entries.

Example

The FreeC.Pass.DefineDeclPass would add a environment entry whose entryIsPartial flag is set to False for the following function declaration.

  head :: [a] -> a
  head xs = case xs of
    []      -> undefined
    x : xs' -> x
  

This pass recognizes that the right-hand side of the function declaration for head refers to the built-in function undefined. Since undefined is partial, head is also partial. Thus, the environment entry of head is updated such that entryIsPartial is set to True.

Specification

Preconditions

The environment contains entries for all function declarations whose entryIsPartial flag is set to False.

Translation

No modifications are made to the AST.

If a function declaration's right-hand side contains a reference to a function whose entry has been marked as partial (i.e., entryIsPartial is set to True), it is marked as partial as well.

If there are mutually recursive function declaration's and one of them needs to be marked as partial by the rule above, then all need to be marked as partial.

Postconditions

The entryIsPartial flag of every environment entry is set to True if an only if the corresponding function is partial.

Documentation

partialityAnalysisPass :: DependencyAwarePass FuncDecl

A compiler pass that updates the partiality information of function declaration environment entries.