FreeC.Pretty

Description

This module contains functions for pretty printing.

We are using the Pretty type class from the 'wl-pprint-text' package.

Documentation

(<$$>) :: Doc -> Doc -> Doc

(<++>) :: Doc -> Doc -> Doc

(<+>) :: Doc -> Doc -> Doc

(<//>) :: Doc -> Doc -> Doc

(</>) :: Doc -> Doc -> Doc

align :: Doc -> Doc

angles :: Doc -> Doc

backslash :: Doc

beside :: Doc -> Doc -> Doc

bool :: Bool -> Doc

braces :: Doc -> Doc

brackets :: Doc -> Doc

cat :: [Doc] -> Doc

char :: Char -> Doc

colon :: Doc

column :: (Int -> Doc) -> Doc

comma :: Doc

displayB :: SimpleDoc -> Builder

displayIO :: Handle -> SimpleDoc -> IO ()

displayT :: SimpleDoc -> Text

displayTStrict :: SimpleDoc -> Text

dot :: Doc

double :: Double -> Doc

dquote :: Doc

dquotes :: Doc -> Doc

empty :: Doc

enclose :: Doc -> Doc -> Doc -> Doc

encloseSep :: Doc -> Doc -> Doc -> [Doc] -> Doc

equals :: Doc

fill :: Int -> Doc -> Doc

fillBreak :: Int -> Doc -> Doc

fillCat :: [Doc] -> Doc

fillSep :: [Doc] -> Doc

float :: Float -> Doc

group :: Doc -> Doc

hPutDoc :: Handle -> Doc -> IO ()

hang :: Int -> Doc -> Doc

hcat :: [Doc] -> Doc

hsep :: [Doc] -> Doc

indent :: Int -> Doc -> Doc

int :: Int -> Doc

integer :: Integer -> Doc

isEmpty :: Doc -> Bool

langle :: Doc

lbrace :: Doc

lbracket :: Doc

line :: Doc

linebreak :: Doc

list :: [Doc] -> Doc

lparen :: Doc

nest :: Int -> Doc -> Doc

nesting :: (Int -> Doc) -> Doc

parens :: Doc -> Doc

punctuate :: Doc -> [Doc] -> [Doc]

putDoc :: Doc -> IO ()

rangle :: Doc

rational :: Rational -> Doc

rbrace :: Doc

rbracket :: Doc

renderCompact :: Doc -> SimpleDoc

renderOneLine :: Doc -> SimpleDoc

renderPretty :: Float -> Int -> Doc -> SimpleDoc

rparen :: Doc

semi :: Doc

semiBraces :: [Doc] -> Doc

sep :: [Doc] -> Doc

softbreak :: Doc

softline :: Doc

space :: Doc

spacebreak :: Doc

squote :: Doc

squotes :: Doc -> Doc

string :: Text -> Doc

stringStrict :: Text -> Doc

text :: Text -> Doc

textStrict :: Text -> Doc

tupled :: [Doc] -> Doc

vcat :: [Doc] -> Doc

vsep :: [Doc] -> Doc

width :: Doc -> (Int -> Doc) -> Doc

data Doc

class Pretty a where

Minimal complete definition

pretty

Methods

pretty :: a -> Doc

prettyList :: [a] -> Doc

data SimpleDoc

Constructors

SEmpty

SChar Char SimpleDoc

SText !Int64 Builder SimpleDoc

SLine !Int64 SimpleDoc

Pretty printing

prettySeparated :: Pretty a => Doc -> [a] -> Doc

Pretty prints a list of pretty printable values by concatenating their documents with the given separator in between.

prettyMaybe :: (Pretty a, Pretty b) => a -> Maybe b -> Doc

Pretty prints the value contained in the given Maybe value or pretty prints a default value.

There is also a Pretty instance for Maybe that produces the empty document if the value is Nothing.

prettyString :: String -> Doc

Pretty prints a string without automatic newlines if the string does not fit onto the page.

prettyText :: String -> Doc

Pretty prints a string such that long lines that don't fit the page are automatically broken between two words.

prettyLines :: String -> Doc

Pretty prints each line of the given string using prettyText and concatenates the resulting documents vertically.

Trailing lines

data TrailingLine a

A pretty printable value with a trailing newline.

Rendering

renderPretty' :: Pretty a => a -> SimpleDoc

Pretty prints a value with a maximum line length of 120 characters of which 80 are allowed to be non-indentation characters.

Output

putPretty :: Pretty a => a -> IO ()

Prints a pretty printable value to stdout.

putPrettyLn :: Pretty a => a -> IO ()

Prints a pretty printable value to stdout with trailing newline.

hPutPretty :: Pretty a => Handle -> a -> IO ()

Prints a pretty printable value to the given file handle.

hPutPrettyLn :: Pretty a => Handle -> a -> IO ()

Prints a pretty printable value to the given file handle and adds a trailing newline.

writePrettyFile :: Pretty a => FilePath -> a -> IO ()

Writes a pretty printable value to the file located at the given path. There is always a trailing newline at the end of the file.

Conversion

showPretty :: Pretty a => a -> String

Convers a pretty printable value to a string.