FreeC.Test.Expectations

Description

This module contains commonly used expectation setters for the comparison of IR AST nodes.

TODO should we maybe re-export FreeC.Monad.Class.Testable?

Similarity test

shouldBeSimilarTo :: (Similar a, Pretty a) => a -> a -> Expectation

Sets the expectation that the given AST nodes are similar.

shouldNotBeSimilarTo :: (Similar a, Pretty a) => a -> a -> Expectation

Sets the expectation that the given AST nodes are notSimilar.

Pretty printing comparision

prettyShouldBe :: (Pretty a, Pretty b) => a -> b -> Expectation

Pretty prints both values and tests whether the resulting strings are equal modulo white space.