FreeC.Util.Parsec

Description

This module contains utility functions for working with the Parsec library.

Documentation

parsecErrorToMessage :: SrcFileMap -> ParseError -> Message

Converts a ParseError to an error Message.

The error message can quote source code from the given source files.

reportParsecError :: MonadReporter r => SrcFileMap -> ParseError -> r a

Reports a fatal ParseError.

The error message can quote source code from the given source files.

runParsecOrFail

Arguments

:: MonadReporter r

=> SrcFile
The file to parse.

-> [t]
The token stream to parse.

-> Parsec [t] () a
The parser to run.

-> r a

Runs the given parser on the given input and reports a fatal error if there is a parse error.

Orphan instances

ConvertibleSrcSpan SourcePos
Converts a Parsec SourcePos to a SrcSpan.

Instance details

Methods

convertSrcSpan :: SourcePos -> SrcSpan