FreeC.Util.Predicate

Description

This module contains utility functions for logical connectives of boolean predicates.

Documentation

(.&&.) :: (a -> Bool) -> (a -> Bool) -> a -> Bool infixr 3

Combines two predicates to a new predicate whose result is the conjunction of the results of the two given predicates.

(.||.) :: (a -> Bool) -> (a -> Bool) -> a -> Bool infixr 2

Combines two predicates to a new predicate whose result is the disjunction of the results of the two given predicates.

(.||^.) :: Monad m => (a -> m Bool) -> (a -> m Bool) -> a -> m Bool infixr 2

.||. lifted to a monad.