Main

Description

This is the main module for the compiler's command line interface.

Documentation

main :: IO ()

The main function of the compiler.

Runs the compiler application and prints all reported messages to stderr.

compiler :: Application ()

Parses and handles the command line options of the application.

Prints the help message if the --help option or no input file was specified. Otherwise all input files are processed (see convertInputModule). If a fatal message is reported while processing any input file, the compiler will exit. All reported messages will be printed to stderr.

selectFrontend :: Application Frontend

Selects the correct frontend or throws an error if such a frontend does not exist.

selectBackend :: Application Backend

Selects the correct backend or throws an error if such a backend does not exist.

parseInputFile :: (SrcFile -> Application Module) -> FilePath -> Application Module

Parses the given input file with the given parser function.

sortInputModules :: [Module] -> Application [Module]

Sorts the given modules based on their dependencies.

If the module dependencies form a cycle, a fatal error is reported.

convertInputModule :: (Module -> Application String) -> Module -> Application (ModName, String)

Converts the given module with the given converter function.

The resulting string is written to the console or output file.

outputModule :: String -> ModName -> String -> Application ()

Output a module that has been generated from a IR module with the given name.

The generated file has the given file extension.

loadRequiredModules :: Module -> Application ()

Loads the environments of modules imported by the given modules from their environment file.

loadImport :: ImportDecl -> Application ()

Loads the environment of the module imported by the given declaration.

loadModule :: SrcSpan -> ModName -> Application ()

Loads the interface of the module with the given name from the configured import path or base library.

The given source span is used in error messages if the module could not be found. Modules from the base library always take precedence. Allowed file extensions for module interface files are .json (for generated module interfaces) and .toml (for manually maintained module interface files).

loadPrelude :: Application ()

Loads the Prelude module from the base library.

This is the only module that we have to load explicitly, because the import of Prelude can be omitted.