FreeC.Backend.Agda.Syntax

Description

This module exports the original Agda AST to reduce coupling with the Agda libraries.

Identifiers

name :: String -> Name

Creates a (not qualified) Agda variable name from a String.

qname :: [Name] -> Name -> QName

Create a qualified identifier given a local identifier as Name and a list of module Names.

qname' :: Name -> QName

Creates a qualified name using an empty list of module names.

Imports

simpleImport :: QName -> Declaration

Creates an open import declaration for the given qualified name.

open import [modName]

Declarations

moduleDecl :: QName -> [Declaration] -> Declaration

Smart constructor for creating a module containing a list of declarations.

funcSig :: Name -> Expr -> Declaration

Smart constructor for creating function type declarations.

funcSig foo expr = foo : expr

Expressions

intLiteral :: Integer -> Expr

Creates an integer literal.

lambda :: [Name] -> Expr -> Expr

Creates a lambda expression, binding the given names and abstracting the given expression.

λ x y … → [expr]

app :: Expr -> Expr -> Expr

Creates an application AST node. Application is left associative and in in type expressions binds stronger than type arrow. For these cases paren- thesis are added automatically.

e a

ident :: String -> Expr

Helper function for creating expressions from Strings.

Types

fun :: Expr -> Expr -> Expr

A smart constructor for non dependent function types.

pi :: [Name] -> Expr -> Expr

Creates a pi type binding the given names as hidden variables.

pi [α₁, …, αₙ] expr ↦ ∀ {α₁} … {αₙ} → expr