FreeC.Backend.Coq.Analysis.DecreasingArguments

Description

This module contains the termination checker for the intermediate language. The goal is to to find one argument for every recursive function whose size is reduced in every recursive call. If such arguments exist we know that the function will terminate and can be translated to Coq. The termination checker defined in this module is very limited compared to Coq's termination checker. Even if we don't recognize the decreasing arguments, Coq might still be able to tell that the function is decreasing on one of its arguments. However, we cannot rely on Coq's termination entirely, since we need to know the decreasing arguments for the transformation of recursive functions into a recursive helper function and a non-recursive main function (see also FreeC.Backend.Coq.Converter.FuncDecl.Rec.WithHelpers)

Guessing Decreasing Arguments

Given set of mutually recursive function declarations F, the termination checker tries to find a combination of arguments such that each function in the set decreases on the corresponding argument. Since we have to know the decreasing argument of all functions in the set in order to test whether a function decreases on one of its argument, we first enumerate (or guess) all possible combinations and then check whether the combination of decreasing arguments is valid.

Checking Decreasing Arguments

A function f ∈ F declared by

  f x₁ … xᵢ … xₙ = e
  

decreases on its i-th argument if for every recursive call to functions g ∈ F on the right-hand side e of f

@

Documentation

type DecArgIndex = Int

Type for the index of a decreasing argument.

identifyDecArgs :: [FuncDecl] -> Converter [DecArgIndex]

Identifies the decreasing arguments of the given mutually recursive function declarations.

Reports a fatal error message, if the decreasing arguments could not be identified.