FreeC.Backend.Coq.Base

Description

This module contains the Coq identifiers of types, constructors and functions defined in the Base library that accompanies the compiler.

Base Library Import

imports :: Sentence

Import sentence for the Free module from the Base Coq library.

baseLibName :: ModuleIdent

The name of the Base Coq library.

generatedLibName :: ModuleIdent

The name of the Coq library where generated Coq files are placed.

Free Monad

free :: Qualid

The Coq identifier for the Free monad.

shape :: Qualid

The Coq identifier for the Shape argument of the Free monad.

pos :: Qualid

The Coq identifier for the Pos argument of the Free monad.

freePureCon :: Qualid

The Coq identifier for the pure constructor of the Free monad.

freeImpureCon :: Qualid

The Coq identifier for the impure constructor of the Free monad.

freeBind :: Qualid

The Coq identifier for the >>= operator of the Free monad.

freeArgs :: [(Qualid, Term)]

The names and types of the parameters that must be passed to the Free monad. These parameters are added automatically to every defined type and function.

forFree :: Qualid

The Coq identifier for the ForFree property.

Partiality

partial :: Qualid

The Coq identifier for the Partial type class.

partialArg :: Qualid

The Coq identifier for the argument of the Partial type class.

partialUndefined :: Qualid

The identifier for the error term undefined.

partialError :: Qualid

The identifier for the error term error.

Modules

qualifiedSmartConstructorModule :: Ident

The name of the local module, where qualified smart constructor notations are defined.

Sharing

injectable :: Qualid

The Coq identifier for the Injectable type class.

injectableBinder :: Binder

The Coq binder for the Injectable type class.

strategy :: Qualid

The Coq identifier for the Strategy type class.

strategyArg :: Qualid

The Coq identifier for the argument of the Strategy type class.

shareableArgs :: Qualid

The Coq binder for the ShareableArgs type class.

shareableArgsBinder :: Qualid -> Binder

The Coq binder for the ShareableArgs type class with the type variable with the given name.

implicitArg :: Term

The Coq identifier for an implicit argument.

share :: Qualid

The Coq Identifier for the share operator.

Effect Selection

selectExplicitArgs :: Effect -> [Term]

Selects the correct explicit function arguments for the given effect.

selectImplicitArgs :: Effect -> [Term]

Selects the correct implicit function arguments for the given effect.

selectTypedImplicitArgs :: Effect -> Int -> [Term]

Like selectImplicitArgs but the arguments have to be inserted after the type arguments the specified number of times.

selectBinders :: Effect -> [Binder]

Selects the correct binder for the given effect.

selectTypedBinders :: Effect -> [Qualid] -> [Binder]

Like selectBinders but the binders are dependent on the type variables with the given names.

Literal Scopes

integerScope :: Ident

The scope of integer literals.

stringScope :: Ident

The scope of string literals used in error terms.

Tactics

proveInd :: Qualid

The tactic that is needed to prove induction schemes.

Reserved Identifiers

reservedIdents :: [Qualid]

All Coq identifiers that are reserved for the Base library.

This does only include identifiers without corresponding Haskell name.