FreeC.IR.Subterm

Description

This module contains data types and function for working with subterms of expressions and type expressions.

There are also functions for finding the names and types of variables that are bound at a given position in an expression.

Direct children

childTerms :: Subterm a => a -> [a]

Gets the child nodes of the given AST node.

replaceChildTerms :: Subterm a => a -> [a] -> Maybe a

Replaces the child nodes of the given AST node.

Positions

newtype Pos

Describes a position of a subterm within a Haskell expression.

Constructors

Pos [Int]

rootPos :: Pos

Position of the root expression.

consPos :: Int -> Pos -> Pos

Extends an position inside of a child node to a position inside of a parent node.

If pos is the position of a subterm s of an expression or type expression tᵢ, then consPos i pos is the position of the subterm s of a term C t₁ ... tᵢ ... tₙ.

parentPos :: Pos -> Maybe Pos

Gets the parent position or Nothing if the given position is the root position.

allPos :: Subterm a => a -> [Pos]

Gets all valid positions of subterms within the given Haskell expression.

above :: Pos -> Pos -> Bool

Tests whether a position is above another one.

below :: Pos -> Pos -> Bool

Tests whether a position is below another one.

leftOf :: Pos -> Pos -> Bool

Tests whether a position is left of another one.

rightOf :: Pos -> Pos -> Bool

Tests whether a position is right of another one.

Subterms

selectSubterm :: Subterm a => a -> Pos -> Maybe a

Selects a subterm of the given expression at the specified position.

Returns Nothing if there is no such subterm.

replaceSubterm

Arguments

:: Subterm a

=> a
The (type) expression whose subterm to replace.

-> Pos
The position of the subterm.

-> a
The (type) expression to replace the subterm with.

-> Maybe a

Replaces a subterm of the given expression or type expression at the specified position with another expression.

Returns Nothing if there is no such subterm.

replaceSubterms :: Subterm a => a -> [(Pos, a)] -> Maybe a

Replaces all subterms at the given positions with other (type) expressions.

Returns Nothing if any of the subterms could not be replaced

Searching for subterms

findSubtermPos :: Subterm a => (a -> Bool) -> a -> [Pos]

Gets a list of positions for subterms of the given expression that satisfy the provided predicate.

findSubterms :: Subterm a => (a -> Bool) -> a -> [a]

Gets a list of subterms of the given expression that satisfy the provided predicate.

findFirstSubterm :: Subterm a => (a -> Bool) -> a -> Maybe a

Gets the first subterm of the given expression that satisfies the provided predicate.

Return Nothing if there is no such subterm.

Bound variables

boundVarsAt :: Expr -> Pos -> Set QName

Gets the names of variables that are bound by lambda abstractions or variable patterns in case-expressions at the given position of an expression.

Returns the empty set if the position is invalid.

boundVarsWithTypeAt :: Expr -> Pos -> Map QName (Maybe Type)

Like boundVarsAt but also returns the annotated type of then variable pattern.

Returns an empty map if the position is invalid.