FreeC.IR.Syntax

Description

This module contains the abstract syntax tree (AST) for the intermediate representation (IR) of the compiler.

The intermediate language is very similar to the subset of Haskell supported by the compiler. The main goal is to make the transformations on the AST and code generation functions easier to comprehend. The IR does have fewer syntactic constructs than Haskell AST that fewer cases need to be distinguished. For example, there is no explicit representation of infix function applications and no list literals. These kinds of syntactic sugar must be removed by the front end.

An additional goal of this AST is to reduce coupling with the parsing library and source language. Ideally the compiler works with any language whose AST can be transformed into this intermediate representation.

A parser for the intermediate language and a description of its syntax can be found in FreeC.Frontend.IR.Parser. While the intermediate language can be parsed, IR nodes are usually created by parsing another language and converting their AST to the IR AST.

Documentation

data Name

An identifier or a symbolic name.

The constructors of this type do not contain source spans because Names are intended to be comparable. They are used as keys to identify nodes of the dependency graph for example.

Constructors

Ident String
An identifier, e.g. Ident "f" for a function f.

Symbol String
A symbolic name, e.g. Symbol "+" for (+).

identFromName :: Name -> Maybe String

Extracts an identifier from a name. Returns Nothing if the given name is a symbol and not an identifier.

data QName

A qualifiable Name.

Constructors

Qual ModName Name
A qualified Name.

UnQual Name
An unqualified Name.

nameFromQName :: QName -> Name

Extracts the name of a qualifiable name.

identFromQName :: QName -> Maybe String

Extracts an identifier from a qualifiable name.

toQual :: ModName -> QName -> QName

Converts a qualifiable name to a name that is qualified with the given module name.

toUnQual :: QName -> QName

Converts a qualifiable name to an unqualified name.

type TypeVarIdent = String

The name of a type variable.

type ModName = String

The name of a module.

type VarName = QName

The name of a function or build-in operator used in prefix notation, e.g. f x y or (+) n m

type ConName = QName

The name of an constructor used in prefix notation, e.g. (:) x xs.

type TypeConName = QName

The name of a type or type constructor, e.g. Integer or [] a

data DeclIdent

The name of a top-level declaration including location information.

Constructors

DeclIdent

Fields

declIdentSrcSpan :: SrcSpan

declIdentName :: QName

internalIdentChar :: Char

The character that is used to mark internal identifiers.

This is used to generate fresh identifiers that don't conflict with user defined identifiers.

isInternalIdent :: String -> Bool

Tests whether the given identifier was generated for internal use only (i.e., contains internalIdentChar).

isInternalName :: Name -> Bool

Tests whether the given name was generated for internal use only (i.e., it is an identifier that matches isInternalIdent).

isInternalQName :: QName -> Bool

Tests whether the given qualifiable name was generated for internal use only (i.e., the qualified name is internal according to isInternalName).

data Module

A module declaration.

Constructors

Module

Fields

modSrcSpan :: SrcSpan

modName :: ModName

modImports :: [ImportDecl]

modTypeDecls :: [TypeDecl]

modTypeSigs :: [TypeSig]

modPragmas :: [Pragma]

modFuncDecls :: [FuncDecl]

data Comment

A comment.

Comments are collected during parsing. However, the final AST contains no comments. Pragmas (see DecArgPragma) are extracted from comments by the front end.

Constructors

BlockComment
A multi-line comment (i.e., {- ... -}).

Fields

commentSrcSpan :: SrcSpan

commentText :: String

LineComment
A single-line comment (i.e., -- ...).

Fields

commentSrcSpan :: SrcSpan

commentText :: String

customPragmaPrefix :: String

All custom pragmas of the compiler start with FreeC.

data Pragma

Data type for custom {-# FreeC ... #-} pragmas.

Constructors

DecArgPragma
A {-# FreeC function DECREASES ON argument #-} or {-# FreeC function DECREASES ON ARGUMENT index #-} pragma.

Fields

pragmaSrcSpan :: SrcSpan

decArgPragmaFuncName :: QName

decArgPragmaArg :: Either String Int

prettyPragma :: Doc -> Doc

Pretty prints a custom {-# FreeC ... #-} pragma with the given contents.

data ImportDecl

An import declaration.

Constructors

ImportDecl

Fields

importSrcSpan :: SrcSpan

importName :: ModName

data TypeVarDecl

The name of a type variable declaration in the head of a data type or type synonym declaration including location information.

Constructors

TypeVarDecl

Fields

typeVarDeclSrcSpan :: SrcSpan

typeVarDeclIdent :: String

typeVarDeclToType :: TypeVarDecl -> Type

Converts the declaration of a type variable to a type.

typeVarDeclName :: TypeVarDecl -> Name

Gets the name of the type variable declared by the given type variable declaration.

typeVarDeclQName :: TypeVarDecl -> QName

Gets the unqualified name of the type variable declared by the given type variable declaration.

data TypeDecl

A data type or type synonym declaration.

Constructors

DataDecl

Fields

typeDeclSrcSpan :: SrcSpan

typeDeclIdent :: DeclIdent

typeDeclArgs :: [TypeVarDecl]

dataDeclCons :: [ConDecl]

TypeSynDecl

Fields

typeDeclSrcSpan :: SrcSpan

typeDeclIdent :: DeclIdent

typeDeclArgs :: [TypeVarDecl]

typeSynDeclRhs :: Type

typeDeclQName :: TypeDecl -> QName

Gets the qualified name of the given type declaration.

typeDeclName :: TypeDecl -> Name

Gets the unqualified name of the given type declaration.

data ConDecl

A constructor declaration.

Constructors

ConDecl

Fields

conDeclSrcSpan :: SrcSpan

conDeclIdent :: DeclIdent

conDeclFields :: [Type]

conDeclQName :: ConDecl -> QName

Gets the qualified name of the given constructor declaration.

conDeclName :: ConDecl -> Name

Gets the unqualified name of the given constructor declaration.

data TypeSig

A type signature of one or more function declarations.

Constructors

TypeSig

Fields

typeSigSrcSpan :: SrcSpan

typeSigDeclIdents :: [DeclIdent]

typeSigTypeSchema :: TypeSchema

data FuncDecl

A function declaration.

Constructors

FuncDecl

Fields

funcDeclSrcSpan :: SrcSpan

A source span that spans the entire function declaration.

funcDeclIdent :: DeclIdent

The name of the function.

funcDeclTypeArgs :: [TypeVarDecl]

The type arguments of the function.

funcDeclArgs :: [VarPat]

The function's argument patterns.

funcDeclRhs :: Expr

The right-hand side of the function declaration.

funcDeclReturnType :: Maybe Type

The return type of the function.

funcDeclQName :: FuncDecl -> QName

Gets the qualified name of the given function declaration.

funcDeclName :: FuncDecl -> Name

Gets the unqualified name of the given function declaration.

funcDeclType :: FuncDecl -> Maybe Type

Gets the type of the given function declaration or Nothing if at least one of the argument or return type is not annotated.

In contrast to funcDeclTypeSchema the function's type arguments are not abstracted away.

funcDeclTypeSchema :: FuncDecl -> Maybe TypeSchema

Gets the type schema of the given function declaration or Nothing if at least one of the argument or the return type is not annotated.

data TypeSchema

A type expression with explicitly introduced type variables.

Constructors

TypeSchema

Fields

typeSchemaSrcSpan :: SrcSpan

typeSchemaArgs :: [TypeVarDecl]

typeSchemaType :: Type

data Type

A type expression.

Build-in types are represented by applications of their type constructors. E.g. the type [a] is represented as TypeApp (TypeCon "[]") (TypeVar "a"). The only exception to this rule is the function type a -> b. It is represented directly as FuncType (TypeVar "a") (TypeVar "b"). The syntax (->) a b is not supported at the moment. This is due to the special role of functions during the translation to Coq.

Constructors

TypeVar
A type variable.

Fields

typeSrcSpan :: SrcSpan

typeVarIdent :: TypeVarIdent

TypeCon
A type constructor.

Fields

typeSrcSpan :: SrcSpan

typeConName :: TypeConName

TypeApp
A type constructor application.

Fields

typeSrcSpan :: SrcSpan

typeAppLhs :: Type

typeAppRhs :: Type

FuncType
A function type.

Fields

typeSrcSpan :: SrcSpan

funcTypeArg :: Type

funcTypeRes :: Type

typeApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> Type
The partially applied type constructor.

-> [Type]
The type arguments to pass to the type constructor.

-> Type

Creates a type constructor application type.

The given source span is inserted into the generated type constructor and every generated type constructor application.

typeConApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> TypeConName
The name of the type constructor to apply.

-> [Type]
The type arguments to pass to the type constructor.

-> Type

Creates a type constructor application type for the constructor with the given name.

The given source span is inserted into the generated type constructor and every generated type constructor application.

funcType :: SrcSpan -> [Type] -> Type -> Type

Creates a function type with the given argument and return types.

splitFuncType :: Type -> Int -> ([Type], Type)

Splits the type of a function or constructor with the given arity into the argument and return types.

This is basically the inverse of funcType.

prettyTypePred :: Int -> Type -> Doc

Pretty prints a type and adds parenthesis if necessary.

The first argument indicates the precedence of the sourrounding context. * 0 - Top level. No parenthesis are neccessary. * 1 - Parenthesis are needed arround function types. * 2 - Parenthesis are also needed arround type constructor applications.

data Expr

An expression.

Constructors

Con
A constructor.

Fields

exprSrcSpan :: SrcSpan

exprConName :: ConName

exprTypeSchema :: Maybe TypeSchema

Var
A function or local variable.

Fields

exprSrcSpan :: SrcSpan

exprVarName :: VarName

exprTypeSchema :: Maybe TypeSchema

App
Function or constructor application.

Fields

exprSrcSpan :: SrcSpan

exprAppLhr :: Expr

exprAppRhs :: Expr

exprTypeSchema :: Maybe TypeSchema

TypeAppExpr
Visible type application.

Fields

exprSrcSpan :: SrcSpan

exprTypeAppLhs :: Expr

exprTypeAppRhs :: Type

exprTypeSchema :: Maybe TypeSchema

If
if expression.

Fields

exprSrcSpan :: SrcSpan

ifExprCond :: Expr

ifExprThen :: Expr

ifExprElse :: Expr

exprTypeSchema :: Maybe TypeSchema

Case
case expression.

Fields

exprSrcSpan :: SrcSpan

caseExprScrutinee :: Expr

caseExprAlts :: [Alt]

exprTypeSchema :: Maybe TypeSchema

Undefined
Error term undefined.

Fields

exprSrcSpan :: SrcSpan

exprTypeSchema :: Maybe TypeSchema

ErrorExpr
Error term error "message".

Fields

exprSrcSpan :: SrcSpan

errorExprMsg :: String

exprTypeSchema :: Maybe TypeSchema

IntLiteral
An integer literal.

Fields

exprSrcSpan :: SrcSpan

intLiteralValue :: Integer

exprTypeSchema :: Maybe TypeSchema

Lambda
A lambda abstraction.

Fields

exprSrcSpan :: SrcSpan

lambdaExprArgs :: [VarPat]

lambdaEprRhs :: Expr

exprTypeSchema :: Maybe TypeSchema

exprType :: Expr -> Maybe Type

Gets the type annotation of the given expression but discards the forall.

Type annotations quantify their type variables usually only if they are as expression type signatures. The type annotations generated during type inference never quantify their type arguments.

untypedCon :: SrcSpan -> ConName -> Expr

Smart constructor for Con without the last argument.

untypedVar :: SrcSpan -> ConName -> Expr

Smart constructor for Var without the last argument.

untypedApp :: SrcSpan -> Expr -> Expr -> Expr

Smart constructor for App without the last argument.

The type annotation is inferred from the callee's type annotation. If it is annotated with a function type, the created expression is annotated with the function type's result type.

untypedTypeAppExpr :: SrcSpan -> Expr -> Type -> Expr

Smart constructor for TypeAppExpr without the last argument.

The type annotation from the expression which is visibly applied is used to annotate the type of this expression.

app :: SrcSpan -> Expr -> [Expr] -> Expr

Creates an expression for applying the given expression to the provided arguments.

The given source span is inserted into the generated function application expressions.

If the given expression's type is annotated with a function type, all generated application nodes are annotated with the corresponding result types. If no more argument types can be split off, the types of the remaining arguments are not annotated.

varApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> VarName
The name of the function to apply.

-> [Expr]
The arguments to pass to the function.

-> Expr

Creates an expression for applying the function with the given name.

The given source span is inserted into the generated function reference and every generated function application.

Since the type of the variable with the given name is not known, no type annotations will be generated.

conApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> ConName
The name of the constructor to apply.

-> [Expr]
The arguments to pass to the constructor.

-> Expr

Creates a data constructor application expression.

The given source span is inserted into the generated constructor reference and every generated constructor application.

Since the type of the constructor with the given name is not known, no type annotations will be generated.

visibleTypeApp :: SrcSpan -> Expr -> [Type] -> Expr

Creates an expression for passing the type arguments of a function or constructor explicitly.

The given source span is inserted into every generated visible type application node.

If the given expression's type is annotated, all generated visible type application nodes are annotated with the same type.

prettyExprPred :: Int -> Expr -> Doc

Pretty prints an expression and adds parenthesis if necessary.

The first argument indicates the precedence of the surrounding context. * 0 - Top level. No parenthesis are necessary. * 1 - Parenthesis are needed around if, case and lambda expressions. * 2 - Parenthesis are also needed around function applications.

data Alt

One alternative of a case expression.

Constructors

Alt

Fields

altSrcSpan :: SrcSpan

altConPat :: ConPat

altVarPats :: [VarPat]

altRhs :: Expr

data ConPat

A constructor pattern used in an alternative of a case expression.

The main purpose of this data type is to add location information to a ConName.

Constructors

ConPat

Fields

conPatSrcSpan :: SrcSpan

conPatName :: ConName

conPatToExpr :: ConPat -> Expr

Converts a constructor pattern to a constructor expression.

data VarPat

A variable pattern used as an argument to a function, lambda abstraction or constructor pattern.

The variable pattern can optionally have a type signature.

Constructors

VarPat

Fields

varPatSrcSpan :: SrcSpan

varPatIdent :: String

varPatType :: Maybe Type

varPatName :: VarPat -> Name

Gets the name of the given variable pattern.

varPatQName :: VarPat -> QName

Gets the qualified name of the given variable pattern.

varPatToExpr :: VarPat -> Expr

Converts a variable pattern to a variable expression.

toVarPat :: String -> VarPat

Converts the given identifier to a variable pattern without type annotation.

preludeModuleName :: ModName

The name of the Prelude module.

TODO once import ... as ... is supported, the Prelude could be renamed by the user.

unitTypeConName :: TypeConName

The name of the unit type constructor.

tupleTypeConName :: Int -> TypeConName

The name of the n-ary tuple type constructor.

listTypeConName :: TypeConName

The name of the list type constructor.

unitConName :: ConName

Name of the unit data constructor.

nilConName :: ConName

The name of the empty list data constructor.

consConName :: ConName

The name of the non empty list data constructor.

tupleConName :: Int -> ConName

The name of the n-ary tuple data constructor.

integerTypeConName :: TypeConName

When inferring the type of IntLiterals this is the type to infer.

boolTypeConName :: TypeConName

When translating if expressions, we annotate the type of the condition with Bool.

negateOpName :: VarName

The unary prefix operator - is translated to the application of the negate function.

errorFuncName :: VarName

The name of the error function.

undefinedFuncName :: VarName

The name of the undefined function.