FreeC.Pass.EtaConversionPass

Description

This module contains a compiler pass that applies η-conversions to expressions such that all function and constructor applications are fully applied.

An η-conversion is the conversion of a partially applied function expression f to a lambda expression \x -> f x that explicitly applies the missing argument.

Motivation

We have to perform η-conversions due to an optimization of the monadic translation of function declarations. An n-ary function declaration of type τ₁ -> … -> τₙ -> τ is translated to a function declaration of type τ₁' -> … -> τₙ' -> τ' where τᵢ' is the translation of type τᵢ. However, an n-ary lambda abstraction of the same type is translated to m (τ₁' -> (τ₂ -> … -> τₙ -> τ)') where m is the target monad. That only the arguments but not the intermediate results of function declarations are lifted to the monad, improves the readability of generated function applications. Intermediate results don't have to be bound since we know that partial applications cannot have an effect. This optimization does not work when functions are applied only partially. Thus, we have to convert partial applications to full applications.

Specification

Preconditions

The arity of all constructors and functions must be known (i.e., there must be corresponding environment entries).

Translation

On the right-hand sides of function declarations, all of the largest sub-expressions of the form

  f α₁ … αₚ e₁ … eₘ
  

where f is the name of an n-ary constructor or function declaration and m < n is replaced by a lambda abstraction

  \x₍ₘ₊₁₎ … xₙ -> f α₁ … αₚ e₁ … eₘ x₍ₘ₊₁₎ … xₙ
  

where x₍ₘ₊₁₎ … xₙ are n-m fresh variables.

Postconditions

All applications of n-ary functions or constructors have at least n arguments.

Documentation

etaConversionPass :: Pass Module

Applies η-conversions to the right-hand sides of all function declarations in the given module until all function and constructor applications are fully applied.

Testing interface

etaConvertExpr :: Expr -> Converter Expr

Applies η-conversions to the given expression and it's sub-expressions until all function and constructor applications are fully applied.